+++ res/config/55cnce_chem.toml
@@ -59,4 +59,4 @@ title = "Roughly 55 Cancri e @ fO2=IW"
     albedo          = true
     mixing_ratios   = true
     animate         = true
-
+    height          = true
