### 2024-09-25 ff237ed
Version 0.8.4. New plot for pressure vs height (new cfg option). Pressure grid uses a smaller topmost level to avoid issues with extrapolation with low opacity compositions. New tests for making plots. Flag for updating (or not) the mixing ratios for radtrans post-saturation check. Fixed loop instability from SIMD when reading SOCRATES radout. Ensures that input stellar spectrum is strictly ascending in wavelength. Minor changes to notebooks
