+++ src/AGNI.jl
@@ -282,35 +282,36 @@ module AGNI
         spfile_name::String    =        cfg["files" ]["input_sf"]
         star_file::String      =        cfg["files" ]["input_star"]
         nlev_centre::Int       =        cfg["execution"]["num_levels"]
-        flag_cnt::Bool         =        cfg["execution" ]["continua"]
-        flag_ray::Bool         =        cfg["execution" ]["rayleigh"]
-        flag_cld::Bool         =        cfg["execution" ]["cloud"]
-        flag_aer::Bool         =        cfg["execution" ]["aerosol"]
-        overlap::Int           =        cfg["execution" ]["overlap_method"]
-        thermo_funct::Bool     =        cfg["execution" ]["thermo_funct"]
-        incl_convect::Bool     =        cfg["execution" ]["convection"]
-        incl_sens::Bool        =        cfg["execution" ]["sensible_heat"]
-        incl_latent::Bool      =        cfg["execution" ]["latent_heat"]
-        sol_type::Int          =        cfg["execution" ]["solution_type"]
-        solvers_cmd::Array{String,1} =  cfg["execution" ]["solvers"]
-        initial_req::Array{String,1} =  cfg["execution" ]["initial_state"]
-        dx_max::Float64        =        cfg["execution" ]["dx_max"]
-        linesearch::Int        =        cfg["execution" ]["linesearch"]
-        easy_start::Bool       =        cfg["execution" ]["easy_start"]
-        conv_atol::Float64     =        cfg["execution" ]["converge_atol"]
-        conv_rtol::Float64     =        cfg["execution" ]["converge_rtol"]
-        max_steps::Int         =        cfg["execution" ]["max_steps"]
-        max_runtime::Float64   =        cfg["execution" ]["max_runtime"]
+        flag_cnt::Bool         =        cfg["execution"]["continua"]
+        flag_ray::Bool         =        cfg["execution"]["rayleigh"]
+        flag_cld::Bool         =        cfg["execution"]["cloud"]
+        flag_aer::Bool         =        cfg["execution"]["aerosol"]
+        overlap::Int           =        cfg["execution"]["overlap_method"]
+        thermo_funct::Bool     =        cfg["execution"]["thermo_funct"]
+        incl_convect::Bool     =        cfg["execution"]["convection"]
+        incl_sens::Bool        =        cfg["execution"]["sensible_heat"]
+        incl_latent::Bool      =        cfg["execution"]["latent_heat"]
+        sol_type::Int          =        cfg["execution"]["solution_type"]
+        solvers_cmd::Array{String,1} =  cfg["execution"]["solvers"]
+        initial_req::Array{String,1} =  cfg["execution"]["initial_state"]
+        dx_max::Float64        =        cfg["execution"]["dx_max"]
+        linesearch::Int        =        cfg["execution"]["linesearch"]
+        easy_start::Bool       =        cfg["execution"]["easy_start"]
+        conv_atol::Float64     =        cfg["execution"]["converge_atol"]
+        conv_rtol::Float64     =        cfg["execution"]["converge_rtol"]
+        max_steps::Int         =        cfg["execution"]["max_steps"]
+        max_runtime::Float64   =        cfg["execution"]["max_runtime"]
 
         #    plotting stuff
-        plt_run::Bool          = cfg["plots"     ]["at_runtime"]
-        plt_tmp::Bool          = cfg["plots"     ]["temperature"]
-        plt_flx::Bool          = cfg["plots"     ]["fluxes"]
[…]
