### 2024-09-26 a5c65a8
Uses a constant Kzz value outside of convective regions, rather than zero
