### 2024-09-26 63fc06a
Uses a constant Kzz value outside of convective regions, rather than zero
