### 2024-10-11 436f45e
Testing notebooks and configs
