+++ res/config/pRT.toml
@@ -1,0 +1,64 @@
+# AGNI configuration file
+title = "petitRADTRANS validation tests"
+
+[planet]
+    tmp_surf        = 3000.0
+    instellation    = 0.0
+    albedo_b        = 0.0
+    s0_fact         = 0.25
+    zenith_angle    = 54.74
+    surface_material= "blackbody"
+    albedo_s        = 0.0
+    radius          = 6.37e6
+    gravity         = 9.81
+    flux_int        = 0.0
+    turb_coeff      = 0.0
+    wind_speed      = 2.0
+
+[files]
+    input_sf        = "res/spectral_files/Honeyside/4096/Honeyside.sf"
+    input_star      = "res/stellar_spectra/sun.txt"
+    output_dir      = "out/"
+
+[composition]
+    p_surf          = 10.0
+    p_top           = 1e-2
+    vmr_dict        = { H2O = 1.0 }
+    include_all     = false
+    chemistry       = 0
+    condensates     = []
+
+[execution]
+    clean_output    = true
+    verbosity       = 1
+    max_steps       = 20000
+    max_runtime     = 400
+    num_levels      = 150
+    continua        = true
+    rayleigh        = false
+    cloud           = false
+    aerosol         = false
+    overlap_method  = 2
+    thermo_funct    = false
+    sensible_heat   = false
+    latent_heat     = false
+    convection      = false
+    solution_type   = 1
+    solvers         = []
+    dx_max          = 400.0
+    initial_state   = ["isothermal", "500"]
+    linesearch      = 0
+    easy_start      = false
+    converge_atol   = 1.0e-2
+    converge_rtol   = 1.0e-3
+
+[plots]
+    at_runtime      = true
+    temperature     = true
+    fluxes          = true
+    contribution    = true
+    emission        = true
+    albedo          = true
+    mixing_ratios   = true
+    animate         = false
+    height          = true
