### 2024-10-15 1eeffd7
Version 0.8.5 - added version for constant gravity in height. Minor rework to paper.
