+++ res/config/default.toml
@@ -46,7 +46,8 @@ title = "Default"                       # Name for this configuration file
     cloud           = false                     # Include water cloud radiative properties?
     aerosol         = false                     # Include aerosol radiative properties?
     overlap_method  = 4                         # Method for treating line overlap.
-    thermo_funct    = true                     # Use temperature-dependent thermodynamic properties?
+    thermo_funct    = true                      # Use temperature-dependent thermodynamic properties?
+    gravity_funct   = true                      # Use height-dependent gravity calculation?
     sensible_heat   = false                     # Include sensible heat transport at the surface?
     latent_heat     = false                     # Include heat release from phase change
     convection      = true                      # Include heat transport by convection
