### 2024-10-15 fe8a05e
Version 0.8.5 - added version for constant gravity in height. Minor rework to paper.
