+++ docs/paper/paper.md
@@ -41,16 +41,11 @@ AGNI can be interacted with as a library; it is used in this way within the [Jup
 
 # Statement of need
 
-It is thought that all rocky planets go through a "magma ocean" stage, where their mantles are fully molten [@lichtenberg_review_2024; @salvador_review_2023]. For some planets this may be their permanent state, while for others it is fleeting. This stage allows for rapid exchange of energy and material between their atmospheres and interiors. Since this phase is likely common to many planets -- including Earth and Venus -- it is important that we understand the physical processes involved, and how these processes interact with each other [@schaefer_redox_2017; @maurice_volatile_2024]. Accurate atmosphere models can allow us to connect the theory of these young planets to telescope observations, since it is primarily through their atmospheric properties that were are able to characterise them [@piette_rocky_2023; @perryman_book_2018]. Modelling these young planets also involves facing several poorly contrained (and unconstrained) quantites which govern their atmospheric composition and dynamics [@guimond_mineralogical_2023; @sossi_redox_2020].
+It is thought that all rocky planets go through a "magma ocean" stage, where their mantles are fully molten [@lichtenberg_review_2024; @elkins_ranges_2008; @schaefer_magma_2016]. For some planets this may be their permanent state, while for others it is fleeting. This stage allows for rapid exchange of energy and material between their atmospheres and interiors. Since this phase is likely common to many planets -- including Earth and Venus -- it is important that we understand the physical processes involved, and how these processes interact with each other [@schaefer_redox_2017; @maurice_volatile_2024]. Accurate atmosphere models can allow us to connect the theory of these young planets to telescope observations, since it is primarily through their atmospheric properties that were are able to characterise them [@piette_rocky_2023; @perryman_book_2018]. Modelling these young planets involves facing several poorly contrained quantites which govern their atmospheric composition [@guimond_mineralogical_2023; @sossi_redox_2020]. Recently, combined observations and modelling of exoplanet TRAPPIST-1 b enabled inference of its past climactic history [@zieba_obs_2023]. @hu_cancri_2024 were able to characterise the atmosphere of 55 Cancri e by using an advanced proprietary atmosphere model.
 
-Several theoretical studies have modelled the atmospheres and evolution of these young planets, but all have made several simplifying assumptions. @lichtenberg_vertically_2021 coupled a simple atmosphere model to real-gas radiative transfer and interior evolution codes to simulate magma ocean evolution, but did not account for the possibility of convective stability in their atmospheres. @krissansen_was_2021 made similar assumptions. @selsis_cool_2023 used a pure-steam radiative-convective model to investigate convective stability, finding that it is likely to occur, but did not extend their work to coupled scenarios which explore the secular evolution of these planets or gas mixtures. @zilinskas_observability_2023 used a multi-gas model to simulate the atmospheres of Sub-Neptune and Super-Earth exoplanets an equilibrium, finding that the opacity of various gaseous species (notably SiO) plays a key role in determining the structure and observable properties of exoplanet atmospheres. The demand for realistic atmosphere modelling in the context of secular magma ocean evolution is apparent.
[…]
