+++ .github/workflows/install_and_test.yml
@@ -26,7 +26,7 @@ jobs:
       - name: Setup Julia
         uses: julia-actions/setup-julia@v2
         with:
-          version: '1.10'
+          version: '1.11'
 
       # Setup SOCRATES
       - name: Get SOCRATES
