### 2024-10-15 6c92440
Version 0.8.5 - added version for constant gravity in height. Minor rework to paper.
