+++ docs/paper/paper.bib
@@ -57,6 +57,68 @@
 	pages = {A59},
 }
 
+
+@article{schaefer_magma_2016,
+doi = {10.3847/0004-637X/829/2/63},
+url = {https://dx.doi.org/10.3847/0004-637X/829/2/63},
+year = {2016},
+month = {sep},
+publisher = {The American Astronomical Society},
+volume = {829},
+number = {2},
+pages = {63},
+author = {Laura Schaefer and Robin D. Wordsworth and Zachory Berta-Thompson and Dimitar Sasselov},
+title = {PREDICTIONS OF THE ATMOSPHERIC COMPOSITION OF GJ 1132b},
+journal = {The Astrophysical Journal},
+}
+
+@article{sergeev_socrates_2023,
+AUTHOR = {Sergeev, D. E. and Mayne, N. J. and Bendall, T. and Boutle, I. A. and Brown, A. and Kav\v{c}i\v{c}, I. and Kent, J. and Kohary, K. and Manners, J. and Melvin, T. and Olivier, E. and Ragta, L. K. and Shipway, B. and Wakelin, J. and Wood, N. and Zerroukat, M.},
+TITLE = {Simulations of idealised 3D atmospheric flows on terrestrial planets using LFRic-Atmosphere},
+JOURNAL = {Geoscientific Model Development},
+VOLUME = {16},
+YEAR = {2023},
+NUMBER = {19},
+PAGES = {5601--5626},
+URL = {https://gmd.copernicus.org/articles/16/5601/2023/},
+DOI = {10.5194/gmd-16-5601-2023}
+}
+
+@ARTICLE{zieba_obs_2023,
+       author = {{Zieba}, Sebastian and {Kreidberg}, Laura and {Ducrot}, Elsa and {Gillon}, Micha{\"e}l and {Morley}, Caroline and {Schaefer}, Laura and {Tamburo}, Patrick and {Koll}, Daniel D.~B. and {Lyu}, Xintong and {Acu{\~n}a}, Lorena and {Agol}, Eric and {Iyer}, Aishwarya R. and {Hu}, Renyu and {Lincowski}, Andrew P. and {Meadows}, Victoria S. and {Selsis}, Franck and {Bolmont}, Emeline and {Mandell}, Avi M. and {Suissa}, Gabrielle},
+        title = "{No thick carbon dioxide atmosphere on the rocky exoplanet TRAPPIST-1 c}",
+      journal = {Nature},
+     keywords = {Astrophysics - Earth and Planetary Astrophysics},
+         year = 2023,
+        month = aug,
+       volume = {620},
+       number = {7975},
+        pages = {746-749},
+          doi = {10.1038/s41586-023-06232-z},
+archivePrefix = {arXiv},
+       eprint = {2306.10150},
+ primaryClass = {astro-ph.EP},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2023Natur.620..746Z},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@ARTICLE{hu_cancri_2024,
+       author = {{Hu}, Renyu and {Bello-Arufe}, Aaron and {Zhang}, Michael and {Paragas}, Kimberly and {Zilinskas}, Mantas and {van Buchem}, Christiaan and {Bess}, Michael and {Patel}, Jayshil and {Ito}, Yuichi and {Damiano}, Mario and {Scheucher}, Markus and {Oza}, Apurva V. and {Knutson}, Heather A. and {Miguel}, Yamila and {Dragomir}, Diana and {Brandeker}, Alexis and {Demory}, Brice-Olivier},
+        title = "{A secondary atmosphere on the rocky exoplanet 55 Cancri e}",
[…]
