+++ res/config/hotdry.toml
@@ -42,6 +42,7 @@ title = "Hot and dry"
     aerosol         = false
     overlap_method  = 4
     thermo_funct    = true
+    gravity_funct   = true
     sensible_heat   = true
     latent_heat     = false
     convection      = true
