+++ res/config/mnras_chem.toml
@@ -42,6 +42,7 @@ title = "ltt3780b_v5/case_00021"
     aerosol         = false
     overlap_method  = 4
     thermo_funct    = true
+    gravity_funct   = true
     sensible_heat   = true
     latent_heat     = true
     convection      = true
