+++ res/config/55cnce_chem.toml
@@ -38,6 +38,7 @@ title = "Roughly 55 Cancri e @ fO2=IW"
     aerosol         = false
     overlap_method  = 4
     thermo_funct    = true
+    gravity_funct   = true
     sensible_heat   = false
     latent_heat     = true
     convection      = true

[…]
