### 2024-10-15 9602ddd
Version 0.9.0. Minor paper changes also
