### 2024-10-15 6e5f529
Version 0.9.0. Minor paper changes also
