+++ res/config/lbl_validation.toml
@@ -1,5 +1,5 @@
 # AGNI configuration file
-title = "petitRADTRANS validation tests"
+title = "Line-by-line validation tests"
 
 [planet]
     tmp_surf        = 600.0
@@ -16,16 +16,16 @@ title = "petitRADTRANS validation tests"
     wind_speed      = 2.0
 
 [files]
-    # input_sf        = "res/spectral_files/Honeyside/4096/Honeyside.sf"
-    # input_star      = "res/stellar_spectra/sun.txt"
-    input_sf        = "out/runtime.sf"
-    input_star      = ""
+    input_sf        = "res/spectral_files/Mallard/318/Mallard.sf"
+    input_star      = "res/stellar_spectra/sun.txt"
+    # input_sf        = "out/ds4096.sf"
+    # input_star      = ""
     output_dir      = "out/"
 
 [composition]
     p_surf          = 123.0
     p_top           = 1e-6
-    vmr_dict        = { H2O = 1.0 }
+    vmr_dict        = { H2O = 0.9, CO2=0.1 }
     include_all     = false
     chemistry       = 0
     condensates     = []
@@ -35,14 +35,14 @@ title = "petitRADTRANS validation tests"
     verbosity       = 1
     max_steps       = 20000
     max_runtime     = 400
-    num_levels      = 200
-    continua        = true
+    num_levels      = 180
+    continua        = false
     rayleigh        = false
     cloud           = false
     aerosol         = false
     overlap_method  = 2
-    thermo_funct    = false
-    gravity_funct   = false
+    thermo_funct    = true
+    gravity_funct   = true
     sensible_heat   = false
     latent_heat     = false
     convection      = false
