+++ docs/src/index.md
@@ -1,4 +1,10 @@
-# AGNI
+```@raw html
+    <img src="assets/logo_title.svg" width=32% class="center"/>
+    <p align="center">
+        <b>A radiative-convective model for lava planet atmospheres</b>
+    </p>
+```
+
 A numerical model for the atmospheres of hot rocky (exo)planets. AGNI's
 primary purpose is to simulate the evolving atmospheres of magma ocean planets, while
 ensuring that radiative-convective equilibrium is sufficiently maintained.
