### 2024-10-28 ebc6074
Socrates installer use absolute path
