### 2024-10-29 d0c1fbb
Greybody surface albedo
