### 2024-10-29 17185b8
Greybody surface albedo
