+++ res/config/condense.toml
@@ -7,7 +7,7 @@ title = "Condensation test"
     albedo_b        = 0.0
     s0_fact         = 0.375
     zenith_angle    = 48.19
-    surface_material= "blackbody"
+    surface_material= "greybody"
     albedo_s        = 0.2
     radius          = 6.37e6
     gravity         = 9.81
