+++ res/config/mnras_chem.toml
@@ -7,7 +7,7 @@ title = "ltt3780b_v5/case_00021"
     albedo_b        = 0.0
     s0_fact         = 0.25
     zenith_angle    = 54.74
-    surface_material= "blackbody"
+    surface_material= "greybody"
     albedo_s        = 0.3
     radius          = 8.552843e+6
     gravity         = 1.275026e+1
