+++ res/config/lbl_validation.toml
@@ -7,7 +7,7 @@ title = "Line-by-line validation tests"
     albedo_b        = 0.0
     s0_fact         = 1.0
     zenith_angle    = 0.0
-    surface_material= "blackbody"
+    surface_material= "greybody"
     albedo_s        = 0.0
     radius          = 6.6e6
     gravity         = 20
