+++ src/phys.jl
@@ -238,7 +238,7 @@ module phys
         # volatile elements
         ("H"  , "#0000aa"),
         ("C"  , "#ff0000"),
-        ("O"  , "#00dd00"),
+        ("O"  , "#00cc00"),
         ("N"  , "#ffaa00"),
         ("S" ,  "#ff22ff"),
         ("P" ,  "#33ccff"),
@@ -248,6 +248,7 @@ module phys
         ("Fe" , "#888888"),
         ("Si" , "#aa2277"),
         ("Mg" , "#996633"),
+        ("Na" , "#1144ff")
     ])
 
     """
