### 2024-10-30 76ce256
Remove netCDFs - moved to  OSF
