### 2024-10-30 40ffff0
Remove netCDFs - moved to  OSF
