### 2024-10-30 d73cb3b
Fix paths for get_something scripts
