### 2024-10-30 d483193
AQUA notebook
