### 2024-10-30 4d2c130
Updated AQUA dict - interpolation uses too much memory
