### 2024-10-30 ab180c1
CFF divided by band width
