### 2024-11-04 eddb5ee
MNRAS plotting
