### 2024-11-05 c51e88e
Store overlap type in atmosphere as string, rather than integer. Updates config options respectively
