### 2024-11-05 f9a55dd
Store overlap type in atmosphere as string, rather than integer. Updates config options respectively
