+++ res/config/mnras_chem.toml
@@ -1,31 +1,31 @@
 # AGNI configuration file
-title = "HD 63433 d (IW+0) at PROTEUS radiative-equilibrium termination"
+title = "TRAPPIST-1 c (IW+5) at PROTEUS solidification"
 
 [planet]
-    tmp_surf        = 2189.153
-    instellation    = 379703.8
+    tmp_surf        = 1676.62
+    instellation    = 10766.43
     albedo_b        = 0.0
     s0_fact         = 0.25
     zenith_angle    = 54.74
     surface_material= "greybody"
     albedo_s        = 0.3
-    radius          = 6797926.0
-    gravity         = 8.625243
+    radius          = 6949977.0
+    gravity         = 10.79357
     flux_int        = 0.0
     turb_coeff      = 0.001
     wind_speed      = 2.0
     skin_k          = 2.0
     skin_d          = 0.01
-    tmp_magma       = 1724.947
+    tmp_magma       = 1676.62
 
 [files]
     input_sf        = "/network/group/aopp/planetary/RTP035_NICHOLLS_PROTEUS/fwl_data/spectral_files/Honeyside/256/Honeyside.sf"
-    input_star      = "/network/group/aopp/planetary/RTP035_NICHOLLS_PROTEUS/article_data/mnras_proteus_agni/hd63433d_oxfu/case_00005/data/0.sflux"
+    input_star      = "/network/group/aopp/planetary/RTP035_NICHOLLS_PROTEUS/article_data/mnras_proteus_agni/trappist1c_oxfu/case_00021/data/88265567.sflux"
     output_dir      = "out/"
 
 [composition]
     p_top           = 1e-5
-    p_dict          = {H2O = 4.32e-2, CO2 = 9.22e-2, H2  = 7.23e-2, CH4 = 2.82e-5, CO  = 7.89e-1, N2  = 3.26e-3, S2  = 1.54e-7, SO2 = 3.09e-6}
+    p_dict          = {H2O=2015.61, CO2=459.33 , H2 =7.14   , CH4=0.00   , CO =5.46   , N2 =1.80   , S2 =33.96  , SO2=137.77}
     include_all     = true
     chemistry       = 1
     condensates     = []
@@ -40,7 +40,7 @@ title = "HD 63433 d (IW+0) at PROTEUS radiative-equilibrium termination"
     rayleigh        = true
     cloud           = false
     aerosol         = false
-    overlap_method  = 4
+    overlap_method  = "ee"
     thermo_funct    = true
     gravity_funct   = true
     sensible_heat   = true
@@ -50,7 +50,7 @@ title = "HD 63433 d (IW+0) at PROTEUS radiative-equilibrium termination"
     solvers         = ["newton"]
     dx_max          = 400.0
     initial_state   = ["loglin", "700"]
-    linesearch      = 0
+    linesearch      = 1
     easy_start      = false
     converge_atol   = 1.0e-2
     converge_rtol   = 1.0e-3
