+++ res/config/condense.toml
@@ -41,7 +41,7 @@ title = "Condensation test"
     rayleigh        = false
     cloud           = false
     aerosol         = false
-    overlap_method  = 4
+    overlap_method  = "ee"
     thermo_funct    = true
     gravity_funct   = true
     sensible_heat   = true
