+++ res/config/default.toml
@@ -45,7 +45,7 @@ title = "Default"                       # Name for this configuration file
     rayleigh        = true                      # Include rayleigh scattering?
     cloud           = false                     # Include water cloud radiative properties?
     aerosol         = false                     # Include aerosol radiative properties?
-    overlap_method  = 4                         # Method for treating line overlap.
+    overlap_method  = "ee"                      # Method for treating line overlap.
     thermo_funct    = true                      # Use temperature-dependent thermodynamic properties?
     gravity_funct   = true                      # Use height-dependent gravity calculation?
     sensible_heat   = false                     # Include sensible heat transport at the surface?
