+++ src/AGNI.jl
@@ -22,7 +22,7 @@ module AGNI
     include("plotting.jl")
     include("energy.jl")
     include("solver.jl")
-    include("realgas.jl")
+    # include("realgas.jl")
 
     # Import src files
     import .phys
@@ -32,7 +32,7 @@ module AGNI
     import .plotting
     import .energy
     import .solver
-    import .realgas
+    # import .realgas
 
     # Export
     # export atmosphere
