### 2024-11-14 bdb416c
Check SOCRATES version
