### 2024-11-19 9a02545
Add config option for rainout
