### 2024-11-19 bc8979b
Add config option for rainout
