+++ res/config/condense.toml
@@ -19,7 +19,7 @@ title = "Condensation test"
     tmp_magma       = 1700.0
 
 [files]
-    input_sf        = "res/spectral_files/nogit/Dayspring48/Dayspring.sf"
+    input_sf        = "res/spectral_files/Dayspring/48/Dayspring.sf"
     input_star      = "res/stellar_spectra/sun.txt"
     output_dir      = "out/"
 
@@ -47,6 +47,7 @@ title = "Condensation test"
     sensible_heat   = true
     latent_heat     = true
     convection      = true
+    rainout         = true
     solution_type   = 3
     solvers         = ["newton"]
     dx_max          = 200.0
