### 2024-11-29 0112e8b
Contribution function in 1D. Do not normalise in file. Version 0.11.1
