### 2024-11-29 455cb45
Contribution function in 1D. Do not normalise in file. Version 0.11.1
