+++ src/atmosphere.jl
@@ -158,7 +158,7 @@ module atmosphere
         surf_flux::Array{Float64, 1}
 
         # Contribution function (to outgoing flux) per-band
-        contfunc_norm::Array{Float64,2}     # LW only, and normalised by maximum value
+        contfunc_band::Array{Float64,2}     # LW only, not normalised
 
         # Sensible heating
         C_d::Float64                        # Turbulent exchange coefficient [dimensionless]
@@ -201,7 +201,10 @@ module atmosphere
         # Heating rate felt at each level [K/day]
         heating_rate::Array{Float64,1}
 
-        # FastChem stuff
+        # Chemistry stuff
+        fastchem_floor::Float64         # Minimum temperature allowed to be sent to FC
+        fastchem_maxiter::Int           # Maximum FC iterations
+        fastchem_xtol::Float64          # FC solver tolerance
         fastchem_flag::Bool             # Fastchem enabled?
         fastchem_work::String           # Path to fastchem working directory
 
@@ -315,8 +318,12 @@ module atmosphere
                     flag_cloud::Bool =          false,
                     thermo_functions::Bool =    true,
                     gravity_functions::Bool =   true,
+
                     use_all_gases::Bool =       false,
-                    fastchem_work::String =     ""
+                    fastchem_work::String =     "",
+                    fastchem_floor::Float64 =   200.0,
+                    fastchem_maxiter::Float64 = 200.0,
+                    fastchem_xtol::Float64 =    1.0e-4,
                     )::Bool
 
         if !isdir(OUT_DIR) && !isfile(OUT_DIR)
@@ -326,7 +333,7 @@ module atmosphere
         @info  "Setting-up a new atmosphere struct"
 
         # Code versions
-        atmos.AGNI_VERSION = "0.11.0"
+        atmos.AGNI_VERSION = "0.11.1"
         atmos.SOCRATES_VERSION = readchomp(joinpath(ENV["RAD_DIR"],"version"))
         @debug "AGNI VERSION = "*atmos.AGNI_VERSION
         @debug "Using SOCRATES at $(ENV["RAD_DIR"])"
@@ -696,6 +703,10 @@ module atmosphere
         else
             @debug "FastChem env variable not set"
         end
+        # other parameters for FC
+        atmos.fastchem_maxiter = fastchem_maxiter
+        atmos.fastchem_floor   = fastchem_floor
+        atmos.fastchem_xtol    = fastchem_xtol
 
         # Record that the parameters are set
         atmos.is_param = true
@@ -709,20 +720,24 @@ module atmosphere
     """
     **Calculate observed radius and bulk density.**
 
-    This is done at the layer probed in transmission.
+    This is done at the layer probed in transmission, which is set at a fixed pressure.
 
     Arguments:
         - `atmos::Atmos_t`          the atmosphere struct instance to be used.
+        - `p_ref::Float64`          observed pressure level [Pa]
 
     Returns:
-        Nothing
+        - `transspec_rho::Float64`  the bulk density observed in transmission
     """
-    function calc_observed_rho!(atmos::atmosphere.Atmos_t)
+    function calc_observed_rho!(atmos::atmosphere.Atmos_t, p_ref::Float64=100.0)::Float64
 
         # transspec_p::Float64            # (INPUT) level probed in transmission [Pa]
         # transspec_r::Float64            # planet radius probed in transmission [m]
         # transspec_m::Float64            # mass [kg] of atmosphere + interior
[…]
