### 2024-11-29 0e12e44
Adjust fc parameters
