### 2024-12-03 472974a
Fastchem max iter default
