### 2024-12-09 0a1d963
Improved cff 1D plot. Warning for cff 2D plot. Store peak pressure of cff (transspec_p) in netcdf. Calc CFF on last solver iteration
