+++ src/plotting.jl
@@ -13,6 +13,7 @@ module plotting
     using LaTeXStrings
     using Printf
     using FFMPEG
+    using Statistics
     import Glob:glob
 
     import ..atmosphere
@@ -20,7 +21,7 @@ module plotting
 
     # Default plotting configuration
     default(fontfamily="sans-serif", framestyle=:box, label=nothing, grid=true,
-            guidefontsize=9, titlefontsize=9)
+            guidefontsize=9, titlefontsize=9, dpi=280)
 
     # Symmetric log
     function _symlog(v::Float64)::Float64
@@ -39,7 +40,6 @@ module plotting
     Plot the temperature-pressure profile.
     """
     function plot_pt(atmos::atmosphere.Atmos_t, fname::String;
-                            dpi::Int=250,
                             size_x::Int=500, size_y::Int=400,
                             incl_magma::Bool=false,
                             title::String="")
@@ -48,8 +48,7 @@ module plotting
         yticks = 10.0 .^ round.(Int,range( log10(ylims[1]), stop=log10(ylims[2]), step=1))
 
         # Create plot
-        plt = plot(ylims=ylims, yticks=yticks, legend=:outertopright,
-                        dpi=dpi, size=(size_x,size_y))
+        plt = plot(ylims=ylims, yticks=yticks, legend=:outertopright, size=(size_x,size_y))
 
         # Plot phase boundary
         if atmos.condense_any
@@ -104,17 +103,14 @@ module plotting
     Plot the height vs pressure profile.
     """
     function plot_height(atmos::atmosphere.Atmos_t, fname::String;
-                                dpi::Int=250,
                                 size_x::Int=500, size_y::Int=400,
-                                incl_magma::Bool=false,
                                 title::String="")
 
         ylims  = (1e-5*atmos.pl[1]/1.5, 1e-5*atmos.pl[end]*1.5)
         yticks = 10.0 .^ round.(Int,range( log10(ylims[1]), stop=log10(ylims[2]), step=1))
 
         # Create plot
-        plt = plot(ylims=ylims, yticks=yticks, legend=:outertopright,
-                        dpi=dpi, size=(size_x,size_y))
+        plt = plot(ylims=ylims, yticks=yticks, legend=:outertopright, size=(size_x,size_y))
 
         # Plot surface
         scatter!(plt, [0.0], [atmos.pl[end]*1e-5], color="brown3", label=L"P_s")
@@ -142,7 +138,6 @@ module plotting
     Plot the cloud mass mixing ratio and area fraction.
     """
     function plot_cloud(atmos::atmosphere.Atmos_t, fname::String;
-                            dpi::Int=250,
                             size_x::Int=500, size_y::Int=400,
                             title::String="")
 
@@ -155,7 +150,7 @@ module plotting
         # Create plot
         plt = plot( xlims=xlims, xticks=xticks,
                     ylims=ylims, yticks=yticks,
-                    legend=:outertopright, dpi=dpi,
+                    legend=:outertopright,
                     size=(size_x,size_y))
 
         # Temperature profile for reference
@@ -186,7 +181,6 @@ module plotting
     Plot the VMRs of the atmosphere at each cell-centre location.
     """
     function plot_vmr(atmos::atmosphere.Atmos_t, fname::String;
-                            dpi::Int=250,
                             size_x::Int=500, size_y::Int=400)
 
         # X-axis minimum allowed left-hand-side limit (log units)
[…]
