### 2024-12-09 81b749e
Improved cff 1D plot. Warning for cff 2D plot. Store peak pressure of cff (transspec_p) in netcdf. Calc CFF on last solver iteration
