### 2024-12-17 ffb9f73
Update docs with citation. Update JOSS paper with citation. Version 1.0.0
