+++ docs/paper/paper.bib
@@ -317,6 +317,16 @@ archivePrefix = {arXiv},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
+@ARTICLE{nicholls_convection_2024,
+       author = {Harrison Nicholls and Raymond Pierrehumbert and Tim Lichtenberg and Laurent Soucasses and Stef Smeets},
+        title = "{Magma Ocean, Water, and the Early Atmosphere of Venus}",
+      journal = {Monthly Notices of the Royal Astronomical Society},
+         year = 2024,
+        month = dec,
+          doi = {10.1093/mnras/stae2772},
+		  url = {https://arxiv.org/abs/2412.11987}
+}
+
 @ARTICLE{lichtenberg_review_2024,
        author = {{Lichtenberg}, Tim and {Miguel}, Yamila},
         title = "{Super-Earths and Earth-like Exoplanets}",
