### 2024-12-17 065f64d
Update docs with citation. Update JOSS paper with citation. Version 1.0.0
