+++ docs/src/index.md
@@ -7,7 +7,8 @@
 
 A numerical model for the atmospheres of hot rocky (exo)planets. AGNI's
 primary purpose is to simulate the evolving atmospheres of magma ocean planets, while
-ensuring that radiative-convective equilibrium is sufficiently maintained.
+ensuring that radiative-convective equilibrium is sufficiently maintained. Pronounced as
+_ag-nee_. Named after the fire deity of Hinduism.
 
 Follow [Getting started](@ref) for information on installing the code and
 obtaining results.
@@ -16,6 +17,8 @@ Contact: `harrison[dot]nicholls[at]physics.ox.ac.uk`
 
 GitHub: [https://github.com/nichollsh/AGNI](https://github.com/nichollsh/AGNI)
 
-Pronounced: _ag-nee_. Named after the fire deity of Hinduism.
+If you use AGNI, please cite the following papers:
+* Nicholls et al., (2024) - https://doi.org/10.1093/mnras/stae2772
+* Nicholls et al., (2025) - submitted to JOSS
 
-This software is available under the GPLv3. Copyright (C) 2024 Harrison Nicholls.
+This software is available under the GPLv3. Copyright (C) 2025 Harrison Nicholls.
