+++ docs/paper/paper.bib
@@ -319,7 +319,7 @@ archivePrefix = {arXiv},
 
 @ARTICLE{nicholls_convection_2024,
        author = {Harrison Nicholls and Raymond Pierrehumbert and Tim Lichtenberg and Laurent Soucasses and Stef Smeets},
-        title = "{Magma Ocean, Water, and the Early Atmosphere of Venus}",
+        title = "{Convective shutdown in the atmospheres of lava worlds}",
       journal = {Monthly Notices of the Royal Astronomical Society},
          year = 2024,
         month = dec,
