### 2024-12-17 #7259 @nichollsh
Please note that the methods/theory paper based on AGNI has now been accepted in MNRAS.

DOI: https://doi.org/10.1093/mnras/stae2772
arXiv: https://arxiv.org/abs/2412.11987
