### 2024-12-18 b6bdfb1
Tidied notebooks. Update citation. Verion 1.0.1
