### 2024-12-18 db6b389
Tidied notebooks. Update citation. Verion 1.0.1
