+++ res/config/cloud.toml
@@ -1,0 +1,67 @@
+# AGNI configuration file
+title = "Cloud radiative properties test"
+
+[planet]
+    tmp_surf        = 2200.0
+    instellation    = 2000.0
+    albedo_b        = 0.0
+    s0_fact         = 0.375
+    zenith_angle    = 48.19
+    surface_material= "greybody"
+    albedo_s        = 0.2
+    radius          = 6.37e6
+    gravity         = 9.81
+    flux_int        = 0.0
+
+[files]
+    # input_sf        = "res/spectral_files/Frostflow/256/Frostflow.sf"
+    # input_star      = "res/stellar_spectra/sun.txt"
+    input_sf  = "out/runtime.sf"
+    input_star = ""
+    output_dir      = "out/"
+
+[composition]
+    p_surf          = 2000.0
+    p_top           = 1e-6
+    vmr_dict        = { H2O=0.8, CO2=0.19, N2=0.01}
+    include_all     = false
+    chemistry       = 0
+    condensates     = ["H2O"]
+
+[execution]
+    clean_output    = false
+    verbosity       = 2
+    max_steps       = 300
+    max_runtime     = 400
+    num_levels      = 100
+    continua        = true
+    rayleigh        = false
+    cloud           = false
+    aerosol         = false
+    overlap_method  = "ro"
+    thermo_funct    = true
+    gravity_funct   = true
+    sensible_heat   = false
+    latent_heat     = false
+    convection      = true
+    rainout         = false
+    solution_type   = 3
+    solvers         = []
+    dx_max          = 200.0
+    initial_state   = ["dry","sat","H2O","str","5"]
+    linesearch      = 2
+    easy_start      = true
+    converge_atol   = 0.1
+    converge_rtol   = 1.0e-3
+
+
+[plots]
+    at_runtime      = true
+    temperature     = true
+    fluxes          = true
+    contribution    = true
+    emission        = true
+    albedo          = true
+    mixing_ratios   = true
+    animate         = false
+    height          = true
