### 2024-12-29 e73d564
Change default kzz to zero
