### 2024-12-29 5d65bdf
Change default kzz to zero
