### 2025-01-10 #7259 @nichollsh
Hi @dfm, thank you for considering our article for publication in JOSS.

This manuscript was submitted in September and is still marked as "waitlisted". However, I understand that potential reviewers may be busy with other submissions. Please could you give me an idea as the to the timescale for your review process? Should I expect the review for our paper to begin within the next couple of months, or is a longer timescale expected?
