### 2025-01-20 #7259 @arfon
@nichollsh – could you take a look a [this list of potential reviewers](https://reviewers.joss.theoj.org/lookup) and identify a few people who would be good candidates to review this submission?

I'd also love to hear if any of [these individuals](https://github.com/openjournals/joss-reviews/issues/7259#issuecomment-2548206694) looks suitable from your perspective.
