### 2025-01-23 #7259 @MartianColonist
I am happy to review this paper for JOSS.
