### 2025-01-27 #7259 @arjunsavel
I'd be glad to review this as well!
