### 2025-01-27 #7726 @editorialbot
[REVIEW]: AGNI: A radiative-convective model for lava planet atmospheres
**Submitting author:** @nichollsh (<a href="http://orcid.org/0000-0002-8368-4641">Harrison Nicholls</a>)
**Repository:** https://github.com/nichollsh/AGNI
**Branch with paper.md** (empty if default branch): main
**Version:** v0.8.3
