+++ src/aqua.jl
@@ -6,7 +6,7 @@ if (abspath(PROGRAM_FILE) == @__FILE__)
     error("The file '$thisfile' is not for direct execution")
 end
 
-module realgas
+module aqua
 
     import ..phys
 
