### 2025-01-29 81eedc5
Store VDW coefficients. Wrapper functions for calculating rho for a mixture or single gas. Calculate rho using ideal gas. Calculate rho using VDW - NEEDS TESTING
