+++ src/phys.jl
@@ -48,69 +48,69 @@ module phys
     # Newton's gravitational constant [m3 kg-1 s-2]
     const G_grav::Float64 = 6.67430e-11 # NIST CODATA
 
-    # Molecule mean molecular weight, kg mol-1
-    const lookup_mmw::Dict{String, Float64} = Dict{String, Float64}([
+    # Table of gas molecule mean molecular weights, kg mol-1
+    const _lookup_mmw::Dict{String, Float64} = Dict([
         # molecules
-        ("H2O", 1.801530E-02 ),
-        ("CO2", 4.401000E-02 ),
-        ("O3", 4.799820E-02 ),
-        ("N2O", 4.401280E-02 ),
-        ("CO", 2.801060E-02 ),
-        ("CH4", 1.604300E-02 ),
-        ("O2", 3.199880E-02 ),
-        ("NO", 3.000610E-02 ),
-        ("SO2", 6.406280E-02 ),
-        ("NO2", 4.600550E-02 ),
-        ("NH3", 1.703060E-02 ),
-        ("HNO3", 6.301290E-02 ),
-        ("N2", 2.801340E-02 ),
-        ("TiO", 6.386600E-02 ),
-        ("VO", 6.694090E-02 ),
-        ("H2", 2.015880E-03 ),
-        ("OCS", 6.007500E-02 ),
-        ("FeH", 5.685300E-02 ),
-        ("CrH", 5.300400E-02 ),
-        ("PH3", 3.399758E-02 ),
-        ("C2H2", 2.603730E-02 ),
-        ("HCN", 2.702530E-02 ),
-        ("H2S", 3.408100E-02 ),
-        ("NO3", 6.301280E-02 ),
-        ("N2O5", 1.080104E-01 ),
-        ("HONO", 4.701340E-02 ),
-        ("HO2NO2", 7.901220E-02 ),
-        ("H2O2", 3.401470E-02 ),
-        ("C2H6", 3.006900E-02 ),
-        ("CH3", 1.503450E-02 ),
-        ("H2CO", 3.002600E-02 ),
-        ("HO2", 3.300670E-02 ),
-        ("HDO", 1.902140E-02 ),
-        ("HCl", 3.646100E-02 ),
-        ("HF", 2.000689E-02 ),
+        ("H2O",     1.801530E-02 ),
+        ("CO2",     4.401000E-02 ),
+        ("O3",      4.799820E-02 ),
+        ("N2O",     4.401280E-02 ),
+        ("CO",      2.801060E-02 ),
+        ("CH4",     1.604300E-02 ),
+        ("O2",      3.199880E-02 ),
+        ("NO",      3.000610E-02 ),
+        ("SO2",     6.406280E-02 ),
+        ("NO2",     4.600550E-02 ),
+        ("NH3",     1.703060E-02 ),
+        ("HNO3",    6.301290E-02 ),
+        ("N2",      2.801340E-02 ),
+        ("TiO",     6.386600E-02 ),
+        ("VO",      6.694090E-02 ),
+        ("H2",      2.015880E-03 ),
+        ("OCS",     6.007500E-02 ),
+        ("FeH",     5.685300E-02 ),
+        ("CrH",     5.300400E-02 ),
+        ("PH3",     3.399758E-02 ),
+        ("C2H2",    2.603730E-02 ),
+        ("HCN",     2.702530E-02 ),
+        ("H2S",     3.408100E-02 ),
+        ("NO3",     6.301280E-02 ),
+        ("N2O5",    1.080104E-01 ),
+        ("HONO",    4.701340E-02 ),
+        ("HO2NO2",  7.901220E-02 ),
+        ("H2O2",    3.401470E-02 ),
[…]
