### 2025-02-09 3f0b408
Calculate density using Amagat's law. Support for reading vdw and aqua eos. Version 1.1.0 and test notebooks
