+++ src/phys.jl
@@ -10,7 +10,8 @@ module phys
 
     using NCDatasets
     using LoggingExtras
-    import PCHIPInterpolation:Interpolator
+    import PCHIPInterpolation:Interpolator # for 1D interpolation
+    import Interpolations: interpolate, Gridded, Linear, Flat, extrapolate, Extrapolation
 
     # Sources:
     # - Pierrehumbert (2010)
@@ -174,58 +175,6 @@ module phys
         ("U",   2.380280000e-01),
     ])
 
-    # Table of Van der Waals coefficients
-    const _lookup_vdw::Dict{String,Tuple{Float64,Float64}} = Dict([
-        ("C2H2"  , (4.516   ,  0.0522   )),
-        ("NH3"   , (4.225   ,  0.0371   )),
-        ("Ar"    , (1.355   ,  0.03201  )),
-        ("C4H10" , (14.66   ,  0.1226   )),
-        ("CO2"   , (3.640   ,  0.04267  )),
-        ("CS2"   , (11.77   ,  0.07685  )),
-        ("CO"    , (1.505   ,  0.03985  )),
-        ("CCl4"  , (19.748  ,  0.1281   )),
-        ("Cl"    , (6.579   ,  0.05622  )),
-        ("CH3Cl" , (7.570   ,  0.06483  )),
-        ("C2N2"  , (7.769   ,  0.06901  )),
-        ("C10H22", (52.74   ,  0.3043   )),
-        ("C2H6"  , (5.562   ,  0.0638   )),
-        ("C2H5OH", (12.18   ,  0.08407  )),
-        ("Fl"    , (1.171   ,  0.0290   )),
-        ("He"    , (0.0346  ,  0.0238   )),
-        ("C7H16" , (31.06   ,  0.2049   )),
-        ("C6H14" , (24.71   ,  0.1735   )),
-        ("N2H4"  , (8.46    ,  0.0462   )),
-        ("H2"    , (0.2476  ,  0.02661  )),
-        ("HBr"   , (4.510   ,  0.04431  )),
-        ("HCl"   , (3.716   ,  0.04081  )),
-        ("HCN"   , (11.29   ,  0.0881   )),
-        ("HF"    , (9.565   ,  0.0739   )),
-        ("HI"    , (6.309   ,  0.0530   )),
-        ("H2Se"  , (5.338   ,  0.04637  )),
-        ("H2S"   , (4.490   ,  0.04287  )),
-        ("Kr"    , (2.349   ,  0.03978  )),
-        ("Hg"    , (8.200   ,  0.01696  )),
-        ("CH4"   , (2.253   ,  0.04278  )),
-        ("CH4O"  , (9.649   ,  0.06702  )),
-        ("Ne"    , (0.2135  ,  0.01709  )),
-        ("NO"    , (1.358   ,  0.02789  )),
-        ("N2"    , (1.370   ,  0.0387   )),
-        ("NO2"   , (5.354   ,  0.04424  )),
-        ("NF3"   , (3.58    ,  0.0545   )),
-        ("N2O"   , (3.832   ,  0.04415  )),
-        ("O2"    , (1.382   ,  0.03186  )),
-        ("O3"    , (3.570   ,  0.0487   )),
-        ("C5H12" , (19.26   ,  0.146    )),
-        ("C6H6O" , (22.93   ,  0.1177   )),
-        ("PH3"   , (4.692   ,  0.05156  )),
-        ("SiH4"  , (4.377   ,  0.05786  )),
-        ("SO2"   , (6.803   ,  0.05636  )),
-        ("SF6"   , (7.857   ,  0.0879   )),
-        ("CCl4"  , (20.01   ,  0.1281   )),
[…]
