+++ CITATION.cff
@@ -5,7 +5,7 @@ authors:
   given-names: "Harrison"
   orcid: "https://orcid.org/0000-0002-8368-4641"
 title: "Convective shutdown in the atmospheres of lava worlds"
-version: 1.0.3
+version: 1.1.3
 doi: 10.1093/mnras/stae2772
 date-released: 2024-12-29
 url: "https://doi.org/10.1093/mnras/stae2772"
