+++ test/runtests.jl
@@ -47,7 +47,7 @@ passing = true
 # -------------
 @info " "
 @info "Testing heat capacity functions"
-data_H2O::phys.Gas_t = phys.load_gas("res/thermodynamics/", "H2O", true)
+data_H2O::phys.Gas_t = phys.load_gas("res/thermodynamics/", "H2O", true, false)
 c_expt::Array{Float64, 1} = [4.975, 35.22, 41.27 , 51.20 , 55.74 ]     # Expected values of cp [J mol-1 K-1]
 t_test::Array{Float64, 1} = [10.0,  500.0, 1000.0, 2000.0, 3000.0]     # Tested values of temperature
 c_obs::Array{Float64,1} = zeros(Float64, 5)
