+++ res/config/default.toml
@@ -46,6 +46,7 @@ title = "Default"                       # Name for this configuration file
     cloud           = false                     # Include water cloud radiative properties?
     aerosol         = false                     # Include aerosol radiative properties?
     overlap_method  = "ee"                      # Method for treating line overlap.
+    real_gas        = true                      # Use real gas equation of state where possible?
     thermo_funct    = true                      # Use temperature-dependent thermodynamic properties?
     gravity_funct   = true                      # Use height-dependent gravity calculation?
     sensible_heat   = false                     # Include sensible heat transport at the surface?
