+++ res/config/cloud.toml
@@ -39,6 +39,7 @@ title = "Cloud radiative properties test"
     cloud           = false
     aerosol         = false
     overlap_method  = "ro"
+    real_gas        = false
     thermo_funct    = true
     gravity_funct   = true
     sensible_heat   = false
