### 2025-02-12 e757b64
Load and use real gas EOS. New config option to enable. Using linear interpolation rather than PCHIP throughout.
