+++ src/phys.jl
@@ -10,7 +10,6 @@ module phys
 
     using NCDatasets
     using LoggingExtras
-    import PCHIPInterpolation:Interpolator # for 1D interpolation
     import Interpolations: interpolate, Gridded, Linear, Flat, extrapolate, Extrapolation
 
     # Sources:
@@ -236,17 +235,17 @@ module phys
         no_sat::Bool                # No saturation data
         sat_T::Array{Float64,1}     # Reference temperatures [K]
         sat_P::Array{Float64,1}     # Corresponding saturation pressures [Pa]
-        sat_I::Interpolator         # PCHIP Interpolator
+        sat_I::Extrapolation        # 1D linear interpolator-extrapolator
 
         # Latent heat (enthalpy) of phase change
         lat_T::Array{Float64,1}     # Reference temperatures [K]
         lat_H::Array{Float64,1}     # Corresponding heats [J kg-1]
-        lat_I::Interpolator         # PCHIP Interpolator
+        lat_I::Extrapolation        # 1D linear interpolator-extrapolator
 
         # Specific heat capacity
         cap_T::Array{Float64,1}     # Reference temperatures [K]
         cap_C::Array{Float64,1}     # Corresponding Cp values [J K-1 kg-1]
-        cap_I::Interpolator         # PCHIP Interpolator
+        cap_I::Extrapolation        # 1D linear interpolator-extrapolator
 
         # Thermal conductivity
         kc::Float64                 # Constant conductivity [J K-1 kg-1]
@@ -258,13 +257,13 @@ module phys
         # Which equation of state should be used for this gas?
         eos::EOS
 
-        # EOS original grid axes
+        # EOS original grid (flattened 2D arrays)
         eos_T::Array{Float64,1}     # temperature [K]
-        eos_P::Array{Float64,1}     # pressure [log10 Pa]
-        eos_V::Array{Float64,2}     # density [kg m-3]
+        eos_P::Array{Float64,1}     # log pressure [log10 Pa]
+        eos_ρ::Array{Float64,1}     # density [kg m-3]
 
         # EOS interpolator with constant-value extrapolation
-        eos_I::Extrapolation
+        eos_I::Extrapolation        # 2D linear interpolator-extrapolator
 
         Gas_t() = new()
     end # end gas struct
@@ -272,7 +271,8 @@ module phys
     """
     Load gas data into a new struct
     """
-    function load_gas(thermo_dir::String, formula::String, tmp_dep::Bool)::Gas_t
+    function load_gas(thermo_dir::String, formula::String,
+                            tmp_dep::Bool, real_gas::Bool)::Gas_t
 
         @debug ("Loading data for gas $formula")
 
@@ -293,127 +293,144 @@ module phys
             end
         end
 
+        # Set plotting color and label
+        gas.plot_color = _pretty_color(formula)
+        gas.plot_label = _pretty_name(formula)
+
         # Thermal conductivity
         gas.kc = 0.0
 
         # Fastchem name (to be learned later)
         gas.fastchem_name = "_unknown"
 
-        # Set plotting color and label
-        gas.plot_color = _pretty_color(formula)
-        gas.plot_label = _pretty_name(formula)
+        # Default parameters, assuming we have no data...
+        gas.mmw = _get_mmw(formula)
+        gas.JANAF_name = "_unknown"
+
+        # heat capacity set to zero
+        gas.cap_T = [0.0, fbig]
+        gas.cap_C = [0.0, 0.0]
+
+        # latent heat set to zero
+        gas.lat_T = [0.0, fbig]
+        gas.lat_H = [0.0, 0.0]
+
[…]
