+++ docs/src/assets/refs.bib
@@ -1,16 +1,3 @@
-@article{fritsch_pchip_1984,
-author = {Fritsch, F. N. and Butland, J.},
-title = {A Method for Constructing Local Monotone Piecewise Cubic Interpolants},
-journal = {SIAM Journal on Scientific and Statistical Computing},
-volume = {5},
-number = {2},
-pages = {300-304},
-year = {1984},
-doi = {10.1137/0905021},
-URL = {https://doi.org/10.1137/0905021},
-eprint = {https://doi.org/10.1137/0905021}
-}
-
 @book{Hapke_2012, place={Cambridge}, edition={2}, title={Theory of Reflectance and Emittance Spectroscopy}, publisher={Cambridge University Press}, author={Hapke, Bruce}, year={2012}}
 
 @ARTICLE{hogstrom_karman_1988,
