+++ res/config/l9859d.toml
@@ -43,6 +43,7 @@ title = "L 98-59 d"
     cloud           = false
     aerosol         = false
     overlap_method  = "ee"
+    real_gas        = true 
     thermo_funct    = true
     gravity_funct   = true
     sensible_heat   = true
