### 2025-02-12 582559d
Tidy up code. Move constants into new file. Simplified config parser.
