+++ src/AGNI.jl
@@ -212,14 +212,24 @@ module AGNI
 
         # Read configuration options from dict
         @info "Using configuration '$(cfg["title"])'"
-        #    planet stuff
-        tmp_surf::Float64      = cfg["planet"]["tmp_surf"]
-        instellation::Float64  = cfg["planet"]["instellation"]
-        albedo_b::Float64      = cfg["planet"]["albedo_b"]
-        asf_sf::Float64        = cfg["planet"]["s0_fact"]
+
+        #    planet structure
         radius::Float64        = cfg["planet"]["radius"]
-        zenith::Float64        = cfg["planet"]["zenith_angle"]
-        gravity::Float64       = cfg["planet"]["gravity"]
+        gravity::Float64       = 0.0
+        mass::Float64          = 0.0
+        if haskey(cfg["planet"],"gravity")
+            gravity = cfg["planet"]["gravity"]
+        end
+        if haskey(cfg["planet"],"mass")
+            mass = cfg["planet"]["gravity"]
+            if gravity > 0.0
+                # overspecified
+                @error "Misconfiguration: provide `mass` OR `gravity`, not both"
+                return false
+            end
+            # convert mass to gravity
+            gravity = phys.grav_accel(mass, radius)
+        end
 
         albedo_s::Float64      = 0.0
         surface_mat::String    = cfg["planet"]["surface_material"]
@@ -232,13 +242,11 @@ module AGNI
         end
 
         #    composition stuff
-        p_top::Float64                  = cfg["composition"]["p_top"]
         condensates::Array{String,1}    = cfg["composition"]["condensates"]
         chem_type::Int                  = cfg["composition"]["chemistry"]
         p_surf::Float64                 = 0.0
         mf_dict::Dict{String, Float64}  = Dict{String, Float64}()
         mf_path::String                 = ""
-        use_all_gases::Bool             = cfg["composition"]["include_all"]
         if haskey(cfg["composition"],"p_surf")
             # set composition using VMRs + Psurf
             if haskey(cfg["composition"],"vmr_dict")
@@ -277,23 +285,12 @@ module AGNI
         end
 
         #    solver stuff
-        spfile_name::String    =        cfg["files" ]["input_sf"]
         star_file::String      =        cfg["files" ]["input_star"]
-        nlev_centre::Int       =        cfg["execution"]["num_levels"]
-        flag_cnt::Bool         =        cfg["execution"]["continua"]
-        flag_ray::Bool         =        cfg["execution"]["rayleigh"]
-        flag_cld::Bool         =        cfg["execution"]["cloud"]
-        flag_aer::Bool         =        cfg["execution"]["aerosol"]
-        overlap::String        =        cfg["execution"]["overlap_method"]
-        real_gas::Bool         =        cfg["execution"]["real_gas"]
-        thermo_funct::Bool     =        cfg["execution"]["thermo_funct"]
-        gravity_funct::Bool    =        cfg["execution"]["gravity_funct"]
         incl_convect::Bool     =        cfg["execution"]["convection"]
         incl_sens::Bool        =        cfg["execution"]["sensible_heat"]
         incl_latent::Bool      =        cfg["execution"]["latent_heat"]
         sol_type::Int          =        cfg["execution"]["solution_type"]
         solvers_cmd::Array{String,1} =  cfg["execution"]["solvers"]
-        initial_req::Array{String,1} =  cfg["execution"]["initial_state"]
         dx_max::Float64        =        cfg["execution"]["dx_max"]
         linesearch::Int        =        cfg["execution"]["linesearch"]
         easy_start::Bool       =        cfg["execution"]["easy_start"]
@@ -304,14 +301,7 @@ module AGNI
         rainout::Bool          =        cfg["execution"]["rainout"]
 
         #    plotting stuff
[…]
