+++ res/realgas/_readme.txt
@@ -1,4 +1,0 @@
-These files contain data relating to real-gas equations of state. For water, this means
-AQUA while for other gas mixtures it means using the van der Walls EOS.
-
-You can find the data on OSF: https://osf.io/xzet3/
