+++ res/thermodynamics/standard.txt
@@ -1,51 +1,0 @@
-# These files are included in simulations when `include_all = true` in cfg file.
-# They will otherwise be included automatically if provided in cfg file.
-C2H4
-CO
-H2O
-H2SO4
-N2
-O2
-OH
-SO
-SiO
-CH4
-CO2
-H2
-H2S
-HCN
-NH3
-OCS
-S2
-S8
-SO2
-SiO2
-O3
-N2O
-NO
-NO2
-HNO3
-FeH
-PH3
-C2H2
-NO3
-N2O5
-HONO
-HO2NO2
-H2O2
-C2H6
-CH3
-H2CO
-HO2
-C
-Fe
-FeO
-H
-Mg2
-Mg
-MgO
-N
-O
-Si
-S
-SO
