### 2025-02-12 0739c52
Tidy up code. Move constants into new file. Simplified config parser.
