+++ src/phys.jl
@@ -12,196 +12,11 @@ module phys
     using LoggingExtras
     import Interpolations: interpolate, Gridded, Linear, Flat, extrapolate, Extrapolation
 
-    # Sources:
-    # - Pierrehumbert (2010)
-    # - NIST
-    # - Sources outlined in Thermo files (https://github.com/nichollsh/Thermo)
-
-    # Universal gas constant, J K-1 mol-1
-    const R_gas::Float64 = 8.314462618 # NIST CODATA
-
-    # Stefan-boltzmann constant, W m−2 K−4
-    const sigma::Float64 =  5.670374419e-8 # NIST CODATA
-
-    # Von Karman constant, dimensionless
-    const k_vk::Float64 = 0.40  # Hogstrom 1988
-
-    # Planck's constant, J s
-    const h_pl::Float64 = 6.62607015e-34  # NIST CODATA
-
-    # Boltzmann constant, J K-1
-    const k_B::Float64 = 1.380649e-23 # NIST CODATA
-
-    # Speed of light, m s-1
-    const c_vac::Float64 = 299792458.0 # NIST CODATA
+    include("consts.jl")
+    using .consts
 
     # A large floating point number
     const fbig::Float64 = 1e90
-
-    # List of elements included in the model
-    const elements_list::Array{String,1} = ["H","C","N","O","S","P",
-                                            "Fe","Mg","Si","Ca","Al"]
-
-    # 0 degrees celcius, in kelvin
-    const zero_celcius::Float64 = 273.15
-
-    # Newton's gravitational constant [m3 kg-1 s-2]
-    const G_grav::Float64 = 6.67430e-11 # NIST CODATA
-
-    # Table of gas molecule mean molecular weights, kg mol-1
-    const _lookup_mmw::Dict{String, Float64} = Dict([
-        # molecules
-        ("H2O",     1.801530E-02 ),
-        ("CO2",     4.401000E-02 ),
-        ("O3",      4.799820E-02 ),
-        ("N2O",     4.401280E-02 ),
-        ("CO",      2.801060E-02 ),
-        ("CH4",     1.604300E-02 ),
-        ("O2",      3.199880E-02 ),
-        ("NO",      3.000610E-02 ),
-        ("SO2",     6.406280E-02 ),
-        ("NO2",     4.600550E-02 ),
-        ("NH3",     1.703060E-02 ),
-        ("HNO3",    6.301290E-02 ),
-        ("N2",      2.801340E-02 ),
-        ("TiO",     6.386600E-02 ),
-        ("VO",      6.694090E-02 ),
-        ("H2",      2.015880E-03 ),
-        ("OCS",     6.007500E-02 ),
-        ("FeH",     5.685300E-02 ),
-        ("CrH",     5.300400E-02 ),
-        ("PH3",     3.399758E-02 ),
-        ("C2H2",    2.603730E-02 ),
-        ("HCN",     2.702530E-02 ),
-        ("H2S",     3.408100E-02 ),
-        ("NO3",     6.301280E-02 ),
-        ("N2O5",    1.080104E-01 ),
-        ("HONO",    4.701340E-02 ),
-        ("HO2NO2",  7.901220E-02 ),
-        ("H2O2",    3.401470E-02 ),
-        ("C2H6",    3.006900E-02 ),
-        ("CH3",     1.503450E-02 ),
-        ("H2CO",    3.002600E-02 ),
-        ("HO2",     3.300670E-02 ),
-        ("HDO",     1.902140E-02 ),
-        ("HCl",     3.646100E-02 ),
-        ("HF",      2.000689E-02 ),
-
[…]
