+++ test/runtests.jl
@@ -211,7 +211,7 @@ atmosphere.setup!(atmos, ROOT_DIR, output_dir,
                          toa_heating, 1.0, 0.0, theta,
                          tmp_surf,
                          gravity, radius,
-                         nlev_centre, p_surf, p_top,
+                         300, p_surf, p_top,
                          mf_dict,"",
                          flag_gcontinuum=true,
                          flag_rayleigh=false,
