+++ src/consts.jl
@@ -1,0 +1,217 @@
+# Contains physical constants
+
+# Not for direct execution
+if (abspath(PROGRAM_FILE) == @__FILE__)
+    thisfile = @__FILE__
+    error("The file '$thisfile' is not for direct execution")
+end
+
+module consts
+    # Sources:
+    # - Pierrehumbert (2010)
+    # - NIST
+    # - Sources outlined in Thermo files (https://github.com/nichollsh/Thermo)
+
+    # Universal gas constant, J K-1 mol-1
+    const R_gas::Float64 = 8.314462618 # NIST CODATA
+    export R_gas
+
+    # Stefan-boltzmann constant, W m−2 K−4
+    const sigma::Float64 =  5.670374419e-8 # NIST CODATA
+    export R_gas
+
+    # Von Karman constant, dimensionless
+    const k_vk::Float64 = 0.40  # Hogstrom 1988
+    export k_vk
+
+    # Planck's constant, J s
+    const h_pl::Float64 = 6.62607015e-34  # NIST CODATA
+    export h_pl
+
+    # Boltzmann constant, J K-1
+    const k_B::Float64 = 1.380649e-23 # NIST CODATA
+    export k_B
+
+    # Speed of light, m s-1
+    const c_vac::Float64 = 299792458.0 # NIST CODATA
+    export c_vac
+
+    # 0 degrees celcius, in kelvin
+    const zero_celcius::Float64 = 273.15
+    export zero_celcius
+
+    # Newton's gravitational constant [m3 kg-1 s-2]
+    const G_grav::Float64 = 6.67430e-11 # NIST CODATA
+    export G_grav
+
+    # List of elements included in the model
+    const elems_standard::Array{String,1} = ["H","C","N","O","S","P",
+                                                "Fe","Mg","Si","Ca","Al"]
+    export elems_standard
+
+    # Standard species
+    const gases_standard::Array{String, 1} = [
+        "C2H4", "CO", "H2O", "H2SO4", "N2", "O2", "OH", "SO", "SiO", "CH4", "CO2", "H2",
+        "H2S", "HCN", "NH3", "OCS", "S2", "S8", "SO2", "SiO2", "O3", "N2O", "NO", "NO2",
+        "HNO3", "FeH", "PH3", "C2H2", "NO3", "N2O5", "HONO", "HO2NO2", "H2O2", "C2H6",
+        "CH3", "H2CO", "HO2", "C", "Fe", "FeO", "H", "Mg2", "Mg", "MgO", "N", "O", "Si",
+        "S", "SO",
+    ]
+    export gases_standard
+
+    # Table of gas molecule mean molecular weights, kg mol-1
+    const _lookup_mmw::Dict{String, Float64} = Dict([
+        # molecules
+        ("H2O",     1.801530E-02 ),
+        ("CO2",     4.401000E-02 ),
+        ("O3",      4.799820E-02 ),
+        ("N2O",     4.401280E-02 ),
+        ("CO",      2.801060E-02 ),
+        ("CH4",     1.604300E-02 ),
+        ("O2",      3.199880E-02 ),
+        ("NO",      3.000610E-02 ),
+        ("SO2",     6.406280E-02 ),
+        ("NO2",     4.600550E-02 ),
+        ("NH3",     1.703060E-02 ),
+        ("HNO3",    6.301290E-02 ),
+        ("N2",      2.801340E-02 ),
+        ("TiO",     6.386600E-02 ),
+        ("VO",      6.694090E-02 ),
[…]
