### 2025-02-12 64a6266
Test for ideal gas EOS - works
