### 2025-02-13 979de59
Disable AQUA. Update tests. Print EOS.
