### 2025-02-13 13556d5
Disable AQUA. Update tests. Print EOS.
