+++ test/runtests.jl
@@ -33,13 +33,16 @@ p_top           = 1e-8
 nlev_centre     = 100
 radius          = 1.0e7    # metres
 gravity         = 10.0      # m s-2
+total  = 0
+failed = 0
 
 rm(output_dir,force=true,recursive=true)
 if !isdir(output_dir) && !isfile(output_dir)
     mkdir(output_dir)
 end
 
-passing = true
+
+rtol   = 1e-3
 
 
 # -------------
@@ -51,21 +54,22 @@ ideal_H2O::phys.Gas_t = phys.load_gas("res/thermodynamics/", "H2O", true, false)
 t_test  = [10.0,  500.0, 1000.0, 2000.0, 3000.0]     # Tested values of temperature
 v_expt  = [4.975, 35.22, 41.27 , 51.20 , 55.74 ]     # Expected values of cp [J mol-1 K-1]
 v_obs   = zero(t_test)
-cp_pass = true
+test_pass = true
 for i in 1:5
     v_obs[i] = phys.get_Cp(ideal_H2O, t_test[i]) * ideal_H2O.mmw # get value and convert units
-    if abs(v_expt[i]- v_obs[i])/v_expt[i] > 0.01  # error must be <1%
-        global cp_pass = false
+    if abs(v_expt[i]- v_obs[i])/v_expt[i] > rtol
+        global test_pass = false
     end
 end
 @info "Expected values = $(v_expt) J mol-1 K-1"
 @info "Modelled values = $(v_obs) J mol-1 K-1"
-if cp_pass
+if test_pass
     @info "Pass"
 else
     @warn "Fail"
-    passing = false
+    failed += 1
 end
+total += 1
 @info "--------------------------"
 
 
@@ -73,52 +77,80 @@ end
 # Test ideal gas equation of state
 # -------------
 @info " "
-@info "Testing ideal gas equation of state"
+@info "Testing H2O ideal gas equation of state"
 t_test = [30.0,  200.0, 373.0,  1273.0,  1000.0] # Tested values of temperature [K]
 p_test = [1e0,   1e3,   1e5,     1e7,    1e8]    # Tested values of pressure [Pa]
 v_expt = [7.22235e-5, 1.08335e-2, 5.80886e-1, 1.70204e1, 2.16670e2]  # Expected rho [kg m-3]
 v_obs  = zero(p_test)
-eos_pass = true
+test_pass = true
 for i in 1:5
     v_obs[i] = phys.calc_rho_gas(t_test[i], p_test[i], ideal_H2O)
-    if abs(v_expt[i] - v_obs[i])/v_expt[i] > 0.01  # error must be <1%
-        global eos_pass = false
+    if abs(v_expt[i] - v_obs[i])/v_expt[i] > rtol
+        global test_pass = false
     end
 end
 @info "Expected values = $(v_expt) kg m-3"
 @info "Modelled values = $(v_obs) kg m-3"
-if eos_pass
+if test_pass
     @info "Pass"
 else
     @warn "Fail"
-    passing = false
+    failed += 1
 end
+total += 1
 @info "--------------------------"
-
 
 # -------------
 # Test AQUA equation of state
 # -------------
 @info " "
-@info "Testing AQUA equation of state"
+@info "Testing H2O AQUA equation of state"
 aqua_H2O::phys.Gas_t = phys.load_gas("res/thermodynamics/", "H2O", true, true)
-v_expt = [4.975, 35.22, 41.27 , 51.20 , 55.74 ]  # Expected rho [kg m-3]
+v_expt = [7.22235e-5, 1.08335e-2, 5.80886e-1, 1.70204e1, 2.16670e2] 
 v_obs  = zero(p_test)
-eos_pass = true
+test_pass = true
 for i in 1:5
[…]
