### 2025-02-13 8ac3a5a
Config option solvers as string, not array. Using general form of Hp and grad_ad in MLT function
