+++ res/config/condense.toml
@@ -42,7 +42,7 @@ title = "Condensation test"
     cloud           = false
     aerosol         = false
     overlap_method  = "ee"
-    real_gas        = true 
+    real_gas        = true
     thermo_funct    = true
     gravity_funct   = true
     sensible_heat   = true
@@ -50,7 +50,7 @@ title = "Condensation test"
     convection      = true
     rainout         = true
     solution_type   = 3
-    solvers         = ["newton"]
+    solver          = "newton"
     dx_max          = 200.0
     initial_state   = ["loglin","1000"]
     linesearch      = 2
