+++ res/config/default.toml
@@ -54,7 +54,7 @@ title = "Default"                       # Name for this configuration file
     convection      = true                      # Include heat transport by convection
     rainout         = false                     # Model rainout - phase change impacts gas mixing ratios, not just energy fluxes
     solution_type   = 0                         # Solution type (see wiki).
-    solvers         = []                        # Ordered list of solvers to apply (see wiki).
+    solver          = ""                        # Solver to use (see wiki).
     dx_max          = 200.0                     # Maximum step size [Kelvin], when using nonlinear solvers
     initial_state   = ["dry", "sat", "H2O"]     # Ordered list of requests describing the initial state of the atmosphere (see wiki).
     linesearch      = 0                         # Linesearch method to be used (0: None, 1: Golden, 2: Backtracking)
