+++ res/config/l9859d.toml
@@ -43,7 +43,7 @@ title = "L 98-59 d"
     cloud           = false
     aerosol         = false
     overlap_method  = "ee"
-    real_gas        = true 
+    real_gas        = true
     thermo_funct    = true
     gravity_funct   = true
     sensible_heat   = true
@@ -51,7 +51,7 @@ title = "L 98-59 d"
     convection      = true
     rainout         = false
     solution_type   = 3
-    solvers         = ["newton"]
+    solver          = "newton"
     dx_max          = 200.0
     initial_state   = ["loglin","600"]
     linesearch      = 0
