### 2025-02-13 b3cb743
Config option solvers as string, not array. Using general form of Hp and grad_ad in MLT function
