+++ res/config/cloud.toml
@@ -47,7 +47,7 @@ title = "Cloud radiative properties test"
     convection      = true
     rainout         = false
     solution_type   = 3
-    solvers         = []
+    solver          = ""
     dx_max          = 200.0
     initial_state   = ["dry","sat","H2O","str","5"]
     linesearch      = 2
