+++ res/config/lbl_validation.toml
@@ -49,7 +49,7 @@ title = "Line-by-line validation tests"
     convection      = false
     rainout         = false
     solution_type   = 1
-    solvers         = []
+    solver          = ""
     dx_max          = 400.0
     initial_state   = ["dry", "sat", "H2O"]
     linesearch      = 0
