+++ res/config/hotdry.toml
@@ -41,7 +41,7 @@ title = "Hot and dry"
     cloud           = false
     aerosol         = false
     overlap_method  = "ee"
-    real_gas        = true 
+    real_gas        = false
     thermo_funct    = true
     gravity_funct   = true
     sensible_heat   = true
@@ -49,7 +49,7 @@ title = "Hot and dry"
     convection      = true
     rainout         = false
     solution_type   = 3
-    solvers         = ["newton"]
+    solver          = "newton"
     dx_max          = 400.0
     initial_state   = ["iso","2300"]
     linesearch      = 0
