### 2025-02-13 1d0607a
Add MLT param to consts.jl and refactor variables. Solve hotdry.toml works with VDW EOS
