+++ src/energy.jl
@@ -325,6 +325,14 @@ module energy
     The mixing length is set to asymptotically approach H (for z>>H) or z (for
     z<H) as per Blackadar (1962). Alternatively, it can be set equal to H.
 
+    The scale height is formulated as:
+        `Hp = P / (ρ g)`
+    The adiabatic lapse rate is formulated as:
+        `∇_ad = dln(T)/dln(P) = (P/T)*(dT/dP) = (P/T)*(1/[ρ c_p])`
+    for an ideal gas, this becomes:
+        `∇_ad = R / (μ c_p)`
+
+
     Arguments:
     - `atmos::Atmos_t`          the atmosphere struct instance to be used.
     - `pmin::Float64`           pressure [bar] below which convection is disabled
@@ -332,27 +340,26 @@ module energy
     """
     function convection!(atmos::atmosphere.Atmos_t; pmin::Float64=1.0e-4, mltype::Int=2)
 
-        pmin *= 1.0e5 # convert bar to Pa
-
         # Reset arrays
-        fill!(atmos.flux_cdry, 0.0)
-        fill!(atmos.Kzz,       atmos.Kzzcst)
+        fill!(atmos.mask_c,     false)
+        fill!(atmos.flux_cdry,  0.0)
+        fill!(atmos.Kzz,        atmos.Kzzcst)
 
         # Work variables
-        H::Float64 = 0.0; l::Float64 = 0.0; w::Float64 = 0.0
+        Hp::Float64 = 0.0; λ::Float64 = 0.0; w::Float64 = 0.0
         m1::Float64 = 0.0; m2::Float64 = 0.0; mt::Float64 = 0.0
         grav::Float64 = 0.0; mu::Float64 = 0.0; c_p::Float64 = 0.0; rho::Float64 = 0.0
-        grad_ad::Float64 = 0.0; grad_pr::Float64 = 0.0
+        ∇_ad::Float64 = 0.0; ∇_pr::Float64 = 0.0
 
         # Loop from bottom upwards (over cell-edges)
         for i in range(start=atmos.nlev_l-1, step=-1, stop=2)
 
             # Profile lapse rate: d(ln T)/d(ln P) = (P/T)*(dT/dP)
-            grad_pr = ( log(atmos.tmp[i-1]/atmos.tmp[i]) )/( log(atmos.p[i-1]/atmos.p[i]) )
+            ∇_pr = ( log(atmos.tmp[i-1]/atmos.tmp[i]) )/( log(atmos.p[i-1]/atmos.p[i]) )
 
             # Optionally skip low pressures
-            if atmos.pl[i] <= pmin
-                continue
+            if atmos.pl[i] <= pmin * 1.0e5  # convert bar to Pa
+                break
             end
 
             # Mass weights
@@ -360,53 +367,52 @@ module energy
             m2 = atmos.layer_mass[i]
             mt = m1+m2
 
-            # Normalise
+            # Normalise weights
             m1 = m1/mt
             m2 = m2/mt
 
+            # Properties interpolated to layer edge
             grav = atmos.layer_grav[i] * m2 + atmos.layer_grav[i-1] * m1
             mu   = atmos.layer_μ[i]    * m2 + atmos.layer_μ[i-1]    * m1
             c_p  = atmos.layer_cp[i]   * m2 + atmos.layer_cp[i-1]   * m1
             rho  = atmos.layer_ρ[i]    * m2 + atmos.layer_ρ[i-1]    * m1
 
-            # Dry convective lapse rate
+            # Dry convective lapse rate, and pressure scale height
             if atmos.real_gas
                 # general solution
-                grad_ad = atmos.pl[i] / (atmos.tmpl[i] * rho * c_p)
[…]
