### 2025-02-13 7f20dad
Inform about EOS in NetCDF
