### 2025-02-13 8c0bfe4
Inform about EOS in NetCDF
