### 2025-02-14 84cc9b4
Update readme and configs and tutorial
