+++ res/thermodynamics/_readme.txt
@@ -1,6 +1,7 @@
 These files are generated using the thermodynamics pipeline found in the
 following GitHub repository. References can be found therein. They contain
-heat capacities, enthalpies of phase change, saturation curves, and scalar
-properties formatted as NetCDF files.
+heat capacities, enthalpies of phase change, saturation curves, tabulated 
+equations of state, and scalar properties. They are NetCDF files.
 
 https://github.com/nichollsh/ThermoTools
+
