+++ res/config/hotdry.toml
@@ -41,7 +41,7 @@ title = "Hot and dry"
     cloud           = false
     aerosol         = false
     overlap_method  = "ee"
-    real_gas        = false
+    real_gas        = true
     thermo_funct    = true
     gravity_funct   = true
     sensible_heat   = true
