+++ res/config/structure.toml
@@ -1,0 +1,70 @@
+# AGNI configuration file
+title = "Hot and dry"
+
+[planet]
+    tmp_surf        = 2000.0
+    instellation    = 44000.0
+    albedo_b        = 0.18
+    s0_fact         = 0.6652
+    zenith_angle    = 60.0
+    surface_material= "res/surface_albedos/lunar_marebasalt.dat"
+    albedo_s        = 0.0
+    radius          = 6.37e6
+    gravity         = 9.81
+    flux_int        = 0.0
+    turb_coeff      = 1.0e-2
+    wind_speed      = 10.0
+
+[files]
+    input_sf        = "res/spectral_files/Frostflow/256/Frostflow.sf"
+    input_star      = "res/stellar_spectra/sun.txt"
+    output_dir      = "out/"
+
+[composition]
+    p_surf          = 2700.0
+    p_top           = 1e-5
+    vmr_dict        = { H2O = 1.0}
+    vmr_path        = ""
+    include_all     = false
+    chemistry       = 0
+    condensates     = []
+
+
+[execution]
+    clean_output    = true
+    verbosity       = 1
+    max_steps       = 200
+    max_runtime     = 400
+    num_levels      = 50
+    continua        = true
+    rayleigh        = true
+    cloud           = false
+    aerosol         = false
+    overlap_method  = "ee"
+    real_gas        = false
+    thermo_funct    = true
+    gravity_funct   = true
+    sensible_heat   = true
+    latent_heat     = false
+    convection      = true
+    rainout         = false
+    solution_type   = 1
+    solver          = ""
+    dx_max          = 400.0
+    initial_state   = ["dry","sat","H2O"]
+    linesearch      = 0
+    easy_start      = false
+    converge_atol   = 1.0e-2
+    converge_rtol   = 1.0e-4
+
+
+[plots]
+    at_runtime      = true
+    temperature     = true
+    fluxes          = true
+    contribution    = true
+    emission        = true
+    albedo          = true
+    mixing_ratios   = true
+    animate         = true
+    height          = true
