### 2025-02-18 6f1364b
Tidy up atmosphere.jl
