### 2025-02-18 cb1f302
Tidy up atmosphere.jl
