+++ src/atmosphere.jl
@@ -828,21 +828,8 @@ module atmosphere
             error("Atmosphere parameters have not been set")
         end
 
-        # Reset arrays
-        fill!(atmos.z         ,     0.0)
-        fill!(atmos.zl        ,     0.0)
-        fill!(atmos.layer_grav,     0.0)
-        fill!(atmos.layer_thick,    0.0)
-        fill!(atmos.layer_mass   ,  0.0)
-
-        # Temporary values
-        g1::Float64 = 0.0; p1::Float64 = 0.0; t1::Float64 = 0.0
-        g2::Float64 = 0.0; p2::Float64 = 0.0; t2::Float64 = 0.0
-        dzc::Float64= 0.0; dzl::Float64 = 0.0
-        GMpl::Float64 = atmos.grav_surf * (atmos.rp^2.0)
-        code::Int64 = 0 # 0: ok, 1: blew up, 2: collapsed
-        dz_max::Float64 = 6e8
-        dz_min::Float64 = 1e-20
+        # Status
+        ok::Bool = true
 
         # MMW
         calc_profile_mmw!(atmos)
@@ -853,6 +840,52 @@ module atmosphere
         # Set density at each level
         calc_profile_density!(atmos)
 
+        # Perform hydrostatic integration
+        ok = ok && calc_profile_height!(atmos, ignore_errors=ignore_errors)
+
+        # Pass arrays to SOCRATES
+        atmos.atm.p[1, :]           .= atmos.p[:]
+        atmos.atm.p_level[1, 0:end] .= atmos.pl[:]
+        atmos.atm.r_layer[1,:]      .= atmos.z[:]  .+ atmos.rp
+        atmos.atm.r_level[1,0:end]  .= atmos.zl[:] .+ atmos.rp
+        atmos.atm.mass[1, :]        .= atmos.layer_mass[:]
+        atmos.atm.density[1,:]      .= atmos.layer_ρ[:]
+
+        return ok
+    end
+
+    """
+    **Calculate height and gravity for all layers.**
+
+    Performs hydrostatic integration from the ground upwards.
+    Requires density, temperature, pressure to have already been set.
+
+    Arguments:
+        - `atmos::Atmos_t`          the atmosphere struct instance to be used.
+        - `ignore_errors::Bool`     do not generate errors from hydrostatic integrator.
+
+    Returns:
+        - `ok::Bool`                function result is ok
+    """
+    function calc_profile_height!(atmos::atmosphere.Atmos_t;
+                                    ignore_errors::Bool=false)::Bool
+
+        # Reset arrays
+        fill!(atmos.z         ,   0.0)
+        fill!(atmos.zl        ,   0.0)
+        fill!(atmos.layer_grav,   0.0)
+        fill!(atmos.layer_thick,  0.0)
+        fill!(atmos.layer_mass ,  0.0)
+
+        # Temporary values
+        g1::Float64 = 0.0; p1::Float64 = 0.0; t1::Float64 = 0.0
+        g2::Float64 = 0.0; p2::Float64 = 0.0; t2::Float64 = 0.0
+        dzc::Float64= 0.0; dzl::Float64 = 0.0
+        GMpl::Float64 = atmos.grav_surf * atmos.rp * atmos.rp
+        code::Int64 = 0 # 0: ok, 1: blew up, 2: collapsed
+        dz_max::Float64 = 1e9
+        dz_min::Float64 = 1e-20
+
         # Integrate from bottom upwards
         for i in range(start=atmos.nlev_c, stop=1, step=-1)
 
@@ -862,7 +895,7 @@ module atmosphere
 
             # Integrate from lower edge to centre
             if atmos.gravity_funct
[…]
