+++ src/chemistry.jl
@@ -10,12 +10,267 @@ module chemistry
 
     # System libraries
     using Printf
-    using LinearAlgebra
+    # using LinearAlgebra
     using Logging
-    using LoopVectorization
+    # using LoopVectorization
+    import DelimitedFiles:readdlm
+
 
     # Local files
     import ..atmosphere
     import ..phys
+
+    """
+    **Calculate composition assuming chemical equilibrium at each level.**
+
+    Uses FastChem to calculate the gas composition at each level of the atmosphere.
+    Volatiles are converted to bulk elemental abundances, which are then provided to
+    FastChem alongside the temperature/pressure profile. FastChem is currently called as an
+    executable, which is not optimal.
+
+    This function DOES NOT automatically recalculate layer properties (e.g. mmw, density).
+
+    Arguments:
+    - `atmos::Atmos_t`                  the atmosphere struct instance to be used.
+    - `chem_type::Int`                  chemistry type (see wiki)
+    - `write_cfg::Bool`                 write config and elements
+
+    Returns:
+    - `state::Int`                      fastchem state (0: success, 1: critical_fail, 2: elem_fail, 3: conv_fail, 4: both_fail)
+    """
+    function fastchem_eqm!(atmos::atmosphere.Atmos_t, chem_type::Int, write_cfg::Bool)::Int
+
+        @debug "Running equilibrium chemistry"
+
+        # Return code
+        state::Int = 0
+
+        # Check fastchem enabled
+        if !atmos.fastchem_flag
+            @warn "Fastchem is not enabled but `fastchem_eqm!` was called. Have you set FC_DIR?"
+            return 1
+        end
+
+        # Check minimum temperature
+        if maximum(atmos.tmpl) < atmos.fastchem_floor
+            @warn "Temperature profile is entirely too cold for FastChem. Not doing chemistry."
+            return 1
+        end
+
+        count_elem_nonzero::Int = 0
+
+        # Paths
+        execpath::String = joinpath(atmos.FC_DIR,       "fastchem")             # Executable file
+        confpath::String = joinpath(atmos.fastchem_work,"config.input")         # Configuration by AGNI
+        elempath::String = joinpath(atmos.fastchem_work,"elements.dat")         # Elements by AGNI
+        chempath::String = joinpath(atmos.fastchem_work,"chemistry.dat")        # Chemistry by FastChem
+
+        # Check file exists
+        write_cfg = write_cfg || !isfile(confpath) || !isfile(elempath)
+
+        # Write config, elements
+        if write_cfg
+            # Write config (fastchem is quite particular about the format)
+            open(confpath,"w") do f
+                write(f,"#Atmospheric profile input file \n")
+                write(f,joinpath(atmos.fastchem_work,"pt.dat")*" \n\n")
+
+                type_char = ["g","ce","cr"]
+                write(f,"#Chemistry calculation type (gas phase only = g, equilibrium condensation = ce, rainout condensation = cr) \n")
+                write(f,"$(type_char[chem_type]) \n\n")
+
+                write(f,"#Chemistry output file \n")
+                write(f,joinpath(atmos.fastchem_work,"chemistry.dat")*" "*joinpath(atmos.fastchem_work,"condensates.dat")*" \n\n")
+
+                write(f,"#Monitor output file \n")
+                write(f,joinpath(atmos.fastchem_work,"monitor.dat")*" \n\n")
+
+                write(f,"#FastChem console verbose level (1 - 4); 1 = almost silent, 4 = detailed console output \n")
+                write(f,"1 \n\n")
+
+                write(f,"#Output mixing ratios (MR) or particle number densities (ND, default) \n")
+                write(f,"ND \n\n")
+
[…]
