### 2025-02-18 6f1b31f
Move more into chemistry.jl, and update tutorials to work outside VSCode
