+++ src/atmosphere.jl
@@ -1606,269 +1606,10 @@ module atmosphere
     end  # end of allocate
 
     """
-    **Normalise gas VMRs, keeping condensates unchanged**
-
-    Only acts on a single model level.
-
-    Parameters:
-    - `atmos::atmosphere.Atmos_t`       atmosphere structure
-    - `i::Int`                          level index to normalise
-    """
-    function normalise_vmrs!(atmos::atmosphere.Atmos_t, i::Int)
-        # Work variables
-        x_con::Float64 =    0.0
-        x_dry::Float64 =    0.0
-        x_dry_old::Float64= 0.0
-
-        # Work out total VMR of all condensing gases
-        for c in atmos.condensates
-            if atmos.gas_sat[c][i]
-                x_con += atmos.gas_vmr[c][i]
-            end
-        end
-
-        # Calculate current and target dry VMRs
-        x_dry = 1.0 - x_con
-        x_dry_old = 0.0
-        for g in atmos.gas_names
-            # skip condensing gases, since their VMR is set by saturation
-            if !atmos.gas_sat[g][i]
-                x_dry_old += atmos.gas_vmr[g][i]
-            end
-        end
-
-        # Renormalise VMR to unity, scaling DRY COMPONENTS ONLY
-        for g in atmos.gas_names
-            if !atmos.gas_sat[g][i]
-                atmos.gas_vmr[g][i] *= x_dry / x_dry_old
-            end
-        end
-
-        # Check total VMR at this level
-        x_tot = 0.0
-        for g in atmos.gas_names
-            x_tot += atmos.gas_vmr[g][i]
-        end
-        if abs(x_tot - 1.0) > 1.0e-5
-            @warn @sprintf("Mixing ratios sum to %.6e (level %d)",x_tot,i)
-        end
-
-        return nothing
-    end
-
-    """
-    **Adjust gas VMRs according to saturation and cold-trap requirements**
-
-    Volatiles which are allowed to condense are rained-out at condensing levels
-    until the gas is exactly saturated, not supersaturated. If evaporation is enabled here,
-    it will lead to enhanced mixing ratios at deeper levels as rain is converted back
-    into gas from a liquid state.
-
-    Arguments:
-    - `atmos::Atmos_t`          the atmosphere struct instance to be used.
-    """
-    function handle_saturation!(atmos::atmosphere.Atmos_t)
-
-        # Single gas case does not apply here
-        if atmos.gas_num == 1
-            return
-        end
-
-        # Parameters
-        evap_enabled::Bool =        false   # Enable re-vaporation of rain
-        evap_efficiency::Float64 =  0.5     # Evaporation efficiency
-
-        # Work arrays
-        maxvmr::Dict{String, Float64} = Dict{String, Float64}() # max running VMR for each condensable
-        cond_kg::Dict{String,Float64} = Dict{String, Float64}() # condensed kg/m2 for each condensable
-        x_sat::Float64 = 0.0
-        supcrit::Bool =  false
-
-        # Set maximum value (for cold trapping)
-        for c in atmos.condensates
-            maxvmr[c] = atmos.gas_vmr[c][end]
-        end
-
-        # Reset mixing ratios to surface values
-        # Reset phase change flags
-        # Reset condensation yield values
-        for g in atmos.gas_names
-           fill!(atmos.gas_vmr[g][1:end-1], atmos.gas_vmr[g][end])
-           fill!(atmos.gas_sat[g],          false)
-           fill!(atmos.gas_yield[g],        0.0)
-        end
-
-        # Reset water cloud
-        fill!(atmos.cloud_arr_r, 0.0)
-        fill!(atmos.cloud_arr_l, 0.0)
[…]
