### 2025-02-18 67e7037
Move more into chemistry.jl, and update tutorials to work outside VSCode
