### 2025-02-18 8cf872c
Tidy up calc_layer_props
