+++ docs/src/setup.md
@@ -19,10 +19,9 @@ Follow the steps below in order to setup the code.
 4. Setup SOCRATES by doing either **ONE** of the following...
     - Follow the instructions on the [SOCRATES GitHub](https://github.com/nichollsh/SOCRATES) page
     - Run `source src/get_socrates.sh`
-5. `julia -e 'using Pkg; Pkg.activate("."); Pkg.build()'`
+5. `bash src/get_agni.sh`
 AGNI is now installed as a package into a Julia environment in the AGNI
-directory. This will also have downloaded some basic input data.
-You should run the tests next.
+directory. This will also have downloaded some basic input data, and have run the tests.
 
 !!! tip
     The `src/get_socrates.sh` script automatically adds the radiation code to your
@@ -30,7 +29,7 @@ You should run the tests next.
     This variable must be set whenever AGNI is being used.
 
 ## Testing
-Now try running the tests in your terminal.
+If you want to run the tests manually, simply use the script in the `test/` folder...
 ```bash
 julia ./test/runtests.jl
 ```
