+++ .github/workflows/install_and_test.yml
@@ -22,11 +22,14 @@ jobs:
           sudo apt update
           sudo apt-get install libnetcdff-dev netcdf-bin gfortran gcc
 
-      # Setup AGNI
+      # Setup Julia
       - name: Setup Julia
         uses: julia-actions/setup-julia@v2
         with:
           version: '1.11'
+
+      - name: Cache Julia
+        uses: julia-actions/cache@v2
 
       # Setup SOCRATES
       - name: Get SOCRATES
@@ -65,6 +68,6 @@ jobs:
         run: |
           export RAD_DIR="/home/runner/work/AGNI/AGNI/SOCRATES"
           export LD_LIBRARY_PATH=""
-          ./test/runtests.jl
+          julia --project=. -e 'using Pkg; Pkg.test()'
 
 
