+++ Project.toml
@@ -25,5 +25,4 @@ Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
 TOML = "fa267f1f-6049-4f14-aa54-33bafae1ed76"
 
 [compat]
-Interpolations = "0.15.1"
 julia = ">= 1.10"
