+++ test/runtests.jl
@@ -16,7 +16,8 @@ using AGNI
 
 
 # Prepare
-output_dir    = "out/"
+RES_DIR         = joinpath(ROOT_DIR,"res/")
+OUT_DIR         = joinpath(ROOT_DIR,"out/")
 p_top           = 1e-8
 nlev_centre     = 100
 radius          = 1.0e7    # metres
@@ -24,21 +25,19 @@ gravity         = 10.0      # m s-2
 total  = 0
 failed = 0
 
-rm(output_dir,force=true,recursive=true)
-if !isdir(output_dir) && !isfile(output_dir)
-    mkdir(output_dir)
+rm(OUT_DIR,force=true,recursive=true)
+if !isdir(OUT_DIR) && !isfile(OUT_DIR)
+    mkdir(OUT_DIR)
 end
 
-
 rtol   = 1e-3
-
 
 # -------------
 # Test heat capacity lookup tables
 # -------------
 @info " "
-@info "Testing heat capacity functions"
-ideal_H2O::phys.Gas_t = phys.load_gas("res/thermodynamics/", "H2O", true, false)
+@info "Testing heat capacity functions for H2O"
+ideal_H2O::phys.Gas_t = phys.load_gas("$RES_DIR/thermodynamics/", "H2O", true, false)
 t_test  = [10.0,  500.0, 1000.0, 2000.0, 3000.0]     # Tested values of temperature
 v_expt  = [4.975, 35.22, 41.27 , 51.20 , 55.74 ]     # Expected values of cp [J mol-1 K-1]
 v_obs   = zero(t_test)
@@ -93,7 +92,7 @@ total += 1
 # -------------
 @info " "
 @info "Testing H2O AQUA equation of state"
-aqua_H2O::phys.Gas_t = phys.load_gas("res/thermodynamics/", "H2O", true, true)
+aqua_H2O::phys.Gas_t = phys.load_gas("$RES_DIR/thermodynamics/", "H2O", true, true)
 v_expt = [9.26121571e2, 7.2269991e-3, 4.35193299e-1, 1.7022212e1, 6.68198662e1]
 v_obs  = zero(p_test)
 test_pass = true
@@ -120,7 +119,7 @@ total += 1
 # -------------
 @info " "
 @info "Testing CO2 VdW equation of state"
-vdw_CO2::phys.Gas_t = phys.load_gas("res/thermodynamics/", "CO2", true, true)
+vdw_CO2::phys.Gas_t = phys.load_gas("$RES_DIR/thermodynamics/", "CO2", true, true)
 v_expt = [2.6465333e-5, 1.7664486595e-2, 1.0612271156, 4.12385376e1, 1.47631823e2]
 v_obs  = zero(p_test)
 test_pass = true
@@ -159,11 +158,11 @@ mf_dict         = Dict([
                         ("N2"  , 0.1),
                         ("H2"  , 0.2)
                         ])
-spfile_name   = joinpath(ROOT_DIR,"res/spectral_files/Dayspring/48/Dayspring.sf")
+spfile_name   ="$RES_DIR/spectral_files/Dayspring/48/Dayspring.sf"
 
 # Setup atmosphere
 atmos = atmosphere.Atmos_t()
-atmosphere.setup!(atmos, ROOT_DIR, output_dir,
+atmosphere.setup!(atmos, ROOT_DIR, OUT_DIR,
                          spfile_name,
                          toa_heating, 1.0, 0.0, theta,
                          tmp_surf,
@@ -207,11 +206,11 @@ mf_dict         = Dict([
                         ("H2O" , 0.8),
                         ("CO2" , 0.2),
                         ])
-spfile_name   = joinpath(ROOT_DIR,"res/spectral_files/Dayspring/48/Dayspring.sf")
[…]
