### 2025-02-19 c8834e3
Allow toggle which real gas EOS is used
