### 2025-02-19 ae59f33
Self gravity
