### 2025-02-19 cba3310
Remove grav_functions option - always enabled now. Update docs
