+++ res/config/condense.toml
@@ -44,7 +44,6 @@ title = "Condensation test"
     overlap_method  = "ee"
     real_gas        = true
     thermo_funct    = true
-    gravity_funct   = true
     sensible_heat   = true
     latent_heat     = true
     convection      = true
