+++ res/config/hotdry.toml
@@ -43,7 +43,6 @@ title = "Hot and dry"
     overlap_method  = "ee"
     real_gas        = true
     thermo_funct    = true
-    gravity_funct   = true
     sensible_heat   = true
     latent_heat     = false
     convection      = true
