### 2025-02-19 29c8458
Remove grav_functions option - always enabled now. Update docs
