+++ res/config/55cnce_chem.toml
@@ -39,7 +39,6 @@ title = "Roughly 55 Cancri e @ fO2=IW"
     overlap_method  = "ee"
     real_gas        = true
     thermo_funct    = true
-    gravity_funct   = true
     sensible_heat   = false
     latent_heat     = true
     convection      = true
