### 2025-02-19 8d6d915
Fix issue with plots not respecting default params]
