+++ src/plotting.jl
@@ -20,8 +20,12 @@ module plotting
     import ..phys
 
     # Default plotting configuration
-    default(fontfamily="sans-serif", framestyle=:box, label=nothing, grid=true,
-            guidefontsize=9, titlefontsize=9, dpi=280)
+    const plt_default = Dict(:fontfamily => "sans-serif",
+                             :framestyle => :box,
+                             :grid       => true,
+                             :guidefontsize => 9,
+                             :titlefontsize => 9,
+                             :dpi => 280)
 
     # Symmetric log
     function _symlog(v::Float64)::Float64
@@ -45,10 +49,11 @@ module plotting
                             title::String="")
 
         ylims  = (1e-5*atmos.pl[1]/1.5, 1e-5*atmos.pl[end]*1.5)
-        yticks = 10.0 .^ round.(Int,range( log10(ylims[1]), stop=log10(ylims[2]), step=1))
+        yticks = 10.0 .^ round.(Int,range(log10(ylims[1]), stop=log10(ylims[2]), step=1))
 
         # Create plot
-        plt = plot(ylims=ylims, yticks=yticks, legend=:outertopright, size=(size_x,size_y))
+        plt = plot(ylims=ylims, yticks=yticks, legend=:outertopright,
+                        size=(size_x,size_y); plt_default...)
 
         # Plot phase boundary
         if atmos.condense_any
@@ -100,9 +105,9 @@ module plotting
     end
 
     """
-    Plot the height vs pressure profile.
+    Plot the radius vs pressure profile.
     """
-    function plot_height(atmos::atmosphere.Atmos_t, fname::String;
+    function plot_radius(atmos::atmosphere.Atmos_t, fname::String;
                                 size_x::Int=500, size_y::Int=400,
                                 title::String="")
 
@@ -110,17 +115,18 @@ module plotting
         yticks = 10.0 .^ round.(Int,range( log10(ylims[1]), stop=log10(ylims[2]), step=1))
 
         # Create plot
-        plt = plot(ylims=ylims, yticks=yticks, legend=:outertopright, size=(size_x,size_y))
+        plt = plot(ylims=ylims, yticks=yticks, legend=:outertopright,
+                        size=(size_x,size_y); plt_default...)
 
         # Plot surface
-        scatter!(plt, [0.0], [atmos.pl[end]*1e-5], color="brown3", label=L"P_s")
+        scatter!(plt, [atmos.rp], [atmos.pl[end]*1e-5], color="brown3", label=L"P_s")
 
         # Plot cell-centres and cell-edges
-        scatter!(plt, atmos.z*1e-3,  atmos.p*1e-5,  msa=0.0, msw=0, ms=1.2, shape=:diamond, label="Centres")
-        scatter!(plt, atmos.zl*1e-3, atmos.pl*1e-5, msa=0.0, msw=0, ms=1.2, shape=:diamond, label="Edges")
+        scatter!(plt, atmos.r*1e-3,  atmos.p*1e-5,  msa=0.0, msw=0, ms=1.2, shape=:diamond, label="Centres")
+        scatter!(plt, atmos.rl*1e-3, atmos.pl*1e-5, msa=0.0, msw=0, ms=1.2, shape=:diamond, label="Edges")
 
         # Decorate
-        xlabel!(plt, "Height [km]")
+        xlabel!(plt, "Radius [km]")
         ylabel!(plt, "Pressure [bar]")
         yflip!(plt)
         yaxis!(plt, yscale=:log10)
@@ -150,8 +156,7 @@ module plotting
         # Create plot
[…]
