+++ docs/src/setup.md
@@ -24,7 +24,7 @@ AGNI is now installed as a package into a Julia environment in the AGNI
 directory. This will also have downloaded some basic input data, and have run the tests.
 
 !!! tip
-    The `src/get_socrates.sh` script automatically adds the radiation code to your
+    The `get_socrates.sh` script automatically adds the radiation code to your
     environment with the variable `RAD_DIR`, which points to the SOCRATES installation.
     This variable must be set whenever AGNI is being used.
 
@@ -35,11 +35,19 @@ julia test/runtests.jl
 ```
 This will print information on whether tests passed or failed.
 
-## Coupling with FastChem
-This can be enabled using the configuration file parameter `composition.chem_type`. Of
-course, it is first necessary to setup FastChem, which can be done by running
-`source src/get_fastchem.sh` and then setting the `FC_DIR` environment variable.
+## Updating
+It's important that you keep AGNI up to date, especially if you are using as part of
+the [PROTEUS framework](https://github.com/FormingWorlds/PROTEUS). Use this script to
+automatically pull changes from GitHub and download any required data files.
+```bash
+bash src/get_agni.sh
+```
 
 ## Using the code
 See [Running the model](@ref) for information on using the code.
 See [Troubleshooting](@ref) for troubleshooting advice.
+
+## Coupling with FastChem
+This can be enabled using the configuration file parameter `composition.chem_type`. Of
+course, it is first necessary to setup FastChem, which can be done by running
+`source src/get_fastchem.sh` and then setting the `FC_DIR` environment variable.
