### 2025-02-19 50fcf6e
Fix issue with plots not respecting default params]
