### 2025-02-20 2a0ffec
Warn if thermo out of date
