### 2025-02-20 a086b44
Validate thermodynamic data - date created. Update tests
