### 2025-02-20 ebbd09e
Validate thermodynamic data - date created. Update tests
