+++ res/blobs/_readme.txt
@@ -1,0 +1,6 @@
+These are binary blobs, redistributed according to their licenses.
+
+b2sum-amd64-linux
+    This is the reference Linux implementation of the BLAKE2b hashing algorithm.
+    It's used in AGNI to verify the integrity of the data files.
+    You can get it from the official website here: https://www.blake2.net/
