### 2025-02-20 0827715
Perform validity check on res/ files
