+++ src/atmosphere.jl
@@ -23,7 +23,7 @@ module atmosphere
     import ..spectrum
 
     # Constants
-    const AGNI_VERSION::String   = "1.2.1"
+    const AGNI_VERSION::String   = "1.3.0"
     const HYDROGRAV_STEPS::Int64 = 30
 
     # Contains data pertaining to the atmosphere (fluxes, temperature, etc.)
