### 2025-02-22 4f905b4
Analytical solution as option in setpt
