### 2025-02-22 7576a2b
Analytical solution as option in setpt
