+++ README.md
@@ -41,6 +41,10 @@ If you use AGNI, please cite the following papers:
 * Nicholls et al., (2024b) - https://doi.org/10.1093/mnras/stae2772
 * Nicholls et al., (2025) - in review at JOSS
 
+## Example
+Below is an animated example of AGNI solving for a temperature-pressure profile.
+<video src="docs/src/assets/demo.mp4" width=100%></video>
+
 ## Repository structure
 * `agni.jl`         - The main AGNI executable
 * `LICENSE.txt`     - License for use and re-use
