### 2025-02-27 22e183d
Use ideal gas EOS in condensing regions to avoid weird issues with AQUA and Amagat's law. Update docs. Store density in netCDF files. Version 1.3.1
