+++ CITATION.cff
@@ -1,4 +1,4 @@
-cff-version: 1.3.0
+cff-version: 1.3.1
 message: "If you use this software, please cite it as below."
 authors:
 - family-names: "Nicholls"
