### 2025-03-05 nichollsh/AGNI#96 @nichollsh
Parallel execution
The current CLI allows a user to run AGNI as a "one-shot" simulation. There are functionalities to run the code through a notebook too. However, there's no simple way to run a grid of models for a range of priors or parameters. This would be very useful to include.
