+++ src/solver.jl
@@ -111,7 +111,7 @@ module solver
                             fdw::Float64=3.0e-5, fdc::Bool=true, fdo::Int=2,
                             method::Int=1, ls_method::Int=1, easy_start::Bool=false,
                             ls_increase::Float64=1.08,
-                            detect_plateau::Bool=true, perturb_all::Bool=false,
+                            detect_plateau::Bool=true, perturb_all::Bool=true,
                             modplot::Int=1, save_frames::Bool=true,
                             modprint::Int=1, plot_jacobian::Bool=false,
                             conv_atol::Float64=1.0e-2, conv_rtol::Float64=1.0e-3
@@ -363,7 +363,7 @@ module solver
 
         # Cost function to minimise
         function _cost(_r::Array)
-            return norm(_r, 2)
+            return norm(_r,3)
         end
 
         # Plot current state
