+++ src/atmosphere.jl
@@ -23,7 +23,7 @@ module atmosphere
     import ..spectrum
 
     # Constants
-    const AGNI_VERSION::String   = "1.3.1"
+    const AGNI_VERSION::String   = "1.3.2"
     const HYDROGRAV_STEPS::Int64 = 40
 
     # Contains data pertaining to the atmosphere (fluxes, temperature, etc.)
