+++ CITATION.cff
@@ -1,4 +1,4 @@
-cff-version: 1.3.1
+cff-version: 1.3.2
 message: "If you use this software, please cite it as below."
 authors:
 - family-names: "Nicholls"
