### 2025-03-30 9ad61f2
Use ideal gas C_p when missing thermo data for a particular species
