### 2025-03-30 795e5c8
Use ideal gas C_p when missing thermo data for a particular species
