### 2025-04-02 #7726 @nichollsh
@arfon, thank you for taking on this paper as the supervising editor. It has been a couple of months since the last update (originally submitted in September 2024), so I wanted to check back here to see what the status is of the review.

@arjunsavel and @MartianColonist, please can you let me know if there are any issues in setting the model up. If there are, I can work on the documentation to improve this.
