### 2025-04-07 1e6e3bc
Using Kzz scaling in stratosphere
