### 2025-04-07 5114a29
Using Kzz scaling in stratosphere
