### 2025-04-14 #7726 @arjunsavel
Hi all,

I've completed my look at `AGNI`. The authors have presented an impressive piece of software that neatly handles a wide range of numerically difficult physical regimes (i.e., rainout condensation). It does this while achieving rapid convergence through a clever set of techniques. I may very well end up being a user myself!

My thoughts are summarized below (many of which are minor typos), separated into paper, documentation, and notes from playing around with the code.

**Paper**
- Maybe add a single plot showing representative output?
- I would also reference Anjali Piette's 2023 work on lava world atmospheres, as it seems to be to be on the high-complexity side: https://iopscience.iop.org/article/10.3847/1538-4357/acdef2/meta
- were there any key assumptions that Zilinskas et al. 2023 made?
- quantites which govern —> quantities that govern
- well established FORTRAN code --> well-established FORTRAN code
- this solution which enables --> this solution, which enables
- climactic --> climatic
- evolution, but did not account --> evolution, but they did not account
- an equilibrium --> in equilibrium
- which explore the range --> that explore the range
- due to continuous --> because of continuous
- which conserves energy --> that conserves energy
- that it may be participate --> that it may participate

**Documentation**
- The documentation is clear, pedagogical, and useful.
- On the broader scale: one of the most exciting things about `AGNI` is that it can be readily coupled in the PROTEUS framework. Would it be feasible to have a straightforward demonstration of this coupling capability?

On model description page:
- solar flux —> stellar flux (I believe)
- models as resolving --> models, as resolving
- typo near "As well as the"
- process which occurs --> process that occurs
- the integrate —> the integrated
- Maybe say why you’d prefer one of the different solution types?
-  single-scattering data which varies -->  single-scattering data that varies
- It would be useful to clarify where in the atmosphere mixing length theory is invoked. I.e., do you treat the atmosphere as fully radiative above the altitude at which the Schwarzchild criterion is no longer met?
- Why must the initial atmosphere not be significantly cooler? So that you don’t condense out something? Is it numerically difficult to “uncondense” a species?
- Please define "TKE."
- number of calculations AGNI  --> number of calculations, AGNI
- Why do you need multiple evaluations of r to fill the matrix? This is probably just something I'm missing :)
- What are ca and cr in the residual section?
- Ideally the documentation would have functionality documentation with, e.g., [autodoc](https://sphinx-rtd-trial.readthedocs.io/en/latest/ext/autodoc.html), but I think the schematics suffice.
- What happens when the equilibrium chemistry isn’t installed? Is the atmosphere assumed to have constant abundances?
- Installation was a bit tricky, but I think that’s more of a nature of getting things installed correctly first on my computer than an AGNI issue.

Getting started page
- setup FastChem --> set up FastChem

Running the model page
- read-on —> read on
- Is it composition.chem_type or composition.chemistry?
- A bit more detail on the chemistry would be nice. I.e., are the initial partial pressures set by the input, then broken apart into constituent elements and passed to Fastchem?
- Would be nice to confirm some things in the input file in the docs too
    - That zenith angle is in degrees
    - The instellation units
    - Gravity units
    - Pressure units
    - include_all?

**Notes from running**
- some of the bash scripts didn’t immediately work
  -  e.g., when running the get_fastchem script, I see:  realpath: bash: No such file or directory
usage: dirname string [...]
 - and when running the get_socrates script, I see: git@github.com: Permission denied (publickey).. I'm using bash (zsh) for these tests.
- Cloud broke: [ ERROR ] Spectral file '/Users/arjunsavel/Desktop/research/AGNI/out/runtime.sf' does not exist
    - Try again after coupling fastchem
- I think the condensation run also broke because the maximum time was reached. This *may* be because of the machine I'm using — [FastChem's precision on ARM](https://github.com/NewStrangeWorlds/FastChem/issues/9) is a bit wonky last I checked, and this particularly mattered for lower-temp condensation.
- Sometimes when the user doesn't have some files (like Dayspring), AGNI's suggestion is to use get_data. But get_data doesn’t give all things (like Dayspring). It’s quick to go to the OSF repo and grab the data, so maybe either add more data that can be acquired with get_data or update the error message to include the OSF repo.
- I sometimes get an error: "Spectral files contain no aerosol data" (when running the 55 Cnc TOML). I wouldn't necessarily expect aerosol data to be in spectral files — are these the opacity functions of the clouds? (as opposed to, e.g., particle densities) A bit of description of the spectral files might be useful in the documentation?
- Related: - When “NO_OPACITY” is printed, does that mean it does not contribute spectrally, but the gas does contribute to eg mean molecular weight? Is there any easy way to see what species are in the different spectral files? i.e. SiO?
