### 2025-04-17 131c605
Enable NetCDF compression (6) and update version and docs
