+++ src/save.jl
@@ -17,6 +17,8 @@ module save
     using Printf
     using Dates
 
+    const nc_comp = Dict(:deflatelevel => 6, :shuffle => true, :checksum => :nochecksum)
+
     """
     Write {Pressure, Temperature, Radius} profile to a CSV file
     """
@@ -68,6 +70,10 @@ module save
 
     """
     Write verbose atmosphere data to a NetCDF file
+
+    Arguments
+    - `atmos::Atmos_t`      Atmosphere object
+    - `fname::String`       Filename to write to
     """
     function write_ncdf(atmos::atmosphere.Atmos_t, fname::String)
 
@@ -122,7 +128,7 @@ module save
             defDim(ds, "nchannels", atmos.dimen.nd_channel)  # Number of spectral channels used for calculations
 
             # ----------------------
-            # Scalar quantities
+            # Scalar quantities (not compressed)
             #    Create variables
             var_max_cff_p = defVar(ds, "max_cff_p",     Float64, (), attrib = OrderedDict("units" => "Pa"))     # Pressure level at which CFF is max [Pa]
             var_tmp_surf =  defVar(ds, "tmp_surf",      Float64, (), attrib = OrderedDict("units" => "K"))      # Surface brightness temperature [K]
@@ -136,18 +142,18 @@ module save
             var_tmax =      defVar(ds, "tceiling",      Float64, (), attrib = OrderedDict("units" => "K"))      # Maximum temperature
             var_plrad =     defVar(ds, "planet_radius", Float64, (), attrib = OrderedDict("units" => "m"))      # Value taken for planet radius
             var_gsurf =     defVar(ds, "surf_gravity",  Float64, (), attrib = OrderedDict("units" => "m s-2"))  # Surface gravity
-            var_fray =      defVar(ds, "flag_rayleigh", Char, ())                                               # Includes rayleigh scattering?
-            var_fcon =      defVar(ds, "flag_continuum",Char, ())                                               # Includes continuum absorption?
-            var_fcld =      defVar(ds, "flag_cloud"    ,Char, ())                                               # Includes clouds?
-            var_tfun =      defVar(ds, "thermo_funct"  ,Char, ())                                               # Using temperature-dependent thermodynamic functions
-            var_rgas =      defVar(ds, "real_gas"      ,Char, ())                                               # Using real gas EOS
-            var_ssol =      defVar(ds, "solved"        ,Char, ())                                               # Has a solver been used?
-            var_scon =      defVar(ds, "converged"     ,Char, ())                                               # Did the solver converge?
+            var_fray =      defVar(ds, "flag_rayleigh", Char,    ())                                               # Includes rayleigh scattering?
+            var_fcon =      defVar(ds, "flag_continuum",Char,    ())                                               # Includes continuum absorption?
+            var_fcld =      defVar(ds, "flag_cloud"    ,Char,    ())                                               # Includes clouds?
+            var_tfun =      defVar(ds, "thermo_funct"  ,Char,    ())                                               # Using temperature-dependent thermodynamic functions
[…]
