### 2025-04-17 bb96856
Enable NetCDF compression (6) and update version and docs
