### 2025-04-17 6bd6df1
Change default compression level to 2 - based on benchmarking. Add comments in save.jl
