### 2025-04-17 53cca2f
Change default compression level to 2 - based on benchmarking. Add comments in save.jl
