### 2025-04-17 #7726 @arjunsavel
oh and one more note:

I might suggest implementing the path length correction for high zenith angles discussed in Section 2.4 of [the second HELIOS paper](https://iopscience.iop.org/article/10.3847/1538-3881/ab1084/meta). I could be mistaken, but I don't think this is currently implemented in `AGNI`.

Doing this may be nontrivial given that the correction requires knowledge of the altitude in a given cell, and my impression from a cursory look at this Guillot implementation is that `AGNI` is mostly on a pressure grid when calculating heating rates.

This change wouldn't be necessary for this review, just something I thought I'd put on your radar — especially if you're interested in producing limb TP profiles to post-process for transmission spectroscopy.
