### 2025-04-21 4e9af70
Store sensible heat in NetCDF file
