+++ docs/paper/paper.bib
@@ -240,6 +240,39 @@ archivePrefix = {arXiv},
 	pages = {216},
 }
 
+@article{krissansen_erosion_2024,
+author="Krissansen-Totton, Joshua
+and Wogan, Nicholas
+and Thompson, Maggie
+and Fortney, Jonathan J.",
+title="The erosion of large primary atmospheres typically leaves behind substantial secondary atmospheres on temperate rocky planets",
+journal="Nature Communications",
+year="2024",
+month="Sep",
+day="27",
+volume="15",
+number="1",
+pages="8374",
+issn="2041-1723",
+doi="10.1038/s41467-024-52642-6",
+url="https://doi.org/10.1038/s41467-024-52642-6"
+}
+
+@article{piette_rocky_2023,
+doi = {10.3847/1538-4357/acdef2},
+url = {https://dx.doi.org/10.3847/1538-4357/acdef2},
+year = {2023},
+month = {aug},
+publisher = {The American Astronomical Society},
+volume = {954},
+number = {1},
+pages = {29},
+author = {Piette, Anjali A. A. and Gao, Peter and Brugman, Kara and Shahar, Anat and Lichtenberg, Tim and Miozzi, Francesca and Driscoll, Peter},
+title = {Rocky Planet or Water World? Observability of Low-density Lava World Atmospheres},
+journal = {The Astrophysical Journal},
+abstract = {Super-Earths span a wide range of bulk densities, indicating a diversity in interior conditions beyond that seen in the solar system. In particular, an emerging population of low-density super-Earths may be explained by volatile-rich interiors. Among these, low-density lava worlds have dayside temperatures that are high enough to evaporate their surfaces, providing a unique opportunity to probe their interior compositions and test for the presence of volatiles. In this work, we investigate the atmospheric observability of low-density lava worlds. We use a radiative-convective model to explore the atmospheric structures and emission spectra of these planets, focusing on three case studies with high observability metrics and substellar temperatures spanning ∼1900–2800 K: HD 86226 c, HD 3167 b, and 55 Cnc e. Given the possibility of mixed volatile and silicate interior compositions for these planets, we consider a range of mixed volatile and rock-vapor atmospheric compositions. This includes a range of volatile fractions and three volatile compositions: water-rich (100% H2O), water with CO2 (80% H2O+20% CO2), and a desiccated O-rich scenario (67% O2+33% CO2). We find that spectral features due to H2O, CO2, SiO, and SiO2 are present in the infrared emission spectra as either emission or absorption features, depending on dayside temperature, volatile fraction, and volatile composition. We further simulate JWST secondary-eclipse observations for each of the three case studies, finding that H2O and/or CO2 could be detected with as few as ∼five eclipses. Detecting volatiles in these atmospheres would provide crucial independent evidence that volatile-rich interiors exist among the super-Earth population.}
+}
+
 
 @article{zilinskas_observability_2023,
 	title = {Observability of silicates in volatile atmospheres of super-{Earths} and sub-{Neptunes}},
@@ -301,30 +334,46 @@ archivePrefix = {arXiv},
 }
 
 @ARTICLE{hammond_photometric_2024,
-       author = {{Hammond}, Mark and {Guimond}, Claire Marie and {Lichtenberg}, Tim and {Nicholls}, Harrison and {Fisher}, Chloe and {Luque}, Rafael and {Meier}, Tobias G. and {Taylor}, Jake and {Changeat}, Quentin and {Dang}, Lisa and {Herbort}, Oliver and {Teske}, Johanna},
[…]
