### 2025-04-22 15bb020
Add section in docs for using juliacall. Update default config TOML so that it works with 'basic' data files
