+++ docs/src/usage.md
@@ -66,6 +66,7 @@ Some parameters:
      - `sat`,       `arg` : apply Clausius-Clapeyron saturation curve for the gas `arg`
      - `ncdf`,      `arg` : load profile from the NetCDF file located at `arg`
      - `loglin`,    `arg` : log-linear profile between `tmp_surf` at the bottom and `arg` at the top
+     - `ana`              : use the Guillot ([2010](https://arxiv.org/abs/1006.4702)) analytical temperature solution
 
     For example, setting `initial_state = ["dry", "sat", "H2O", "str", "180"]` will set T(p) to follow the dry adiabat from the surface, the water condensation curve above that, and then to be isothermal at 180 K until the top of the model.
 
@@ -76,9 +77,35 @@ Some parameters:
      - 3 : Equilibrium, with condensation (condensates rained out)
 
 ## Outputs
-Results are optionally plotted and animated, and data will be saved as NetCDF
-or CSV files.
+Results are optionally plotted and animated, and data will be saved as NetCDF or CSV files.
 
-## Python
-It is possible to interact with the model using Python. This is best done with the `juliacall` package from [PythonCall.jl](https://github.com/JuliaPy/PythonCall.jl), and is implemented this way in the [PROTEUS framework](https://github.com/FormingWorlds/PROTEUS).
+## Accessing AGNI from Python
+It is possible to interact with AGNI from Python. This is best done with the `juliacall` package from [PythonCall.jl](https://github.com/JuliaPy/PythonCall.jl).
 
+Coupling with Python is done via `juliacall` within the modular [PROTEUS framework](https://github.com/FormingWorlds/PROTEUS), which couples AGNI self-consistently to models of planetary interior evolution and volatile outgassing. You can see this implemented in [agni.py](https://github.com/FormingWorlds/PROTEUS/blob/main/src/proteus/atmos_clim/agni.py) within the PROTEUS source code.
+
+A skeleton example is given below:
+```python
+
+# Import juliacall
+from juliacall import Main as jl
+
+# Import AGNI
+jl.seval("using Pkg")
+jl.Pkg.activate(AGNI_ROOT_DIR)  # <---- set AGNI_ROOT_DIR to your installation path
+jl.seval("import AGNI")
+jl.AGNI.setup_logging("out.log", 1)
+
+# Setup atmosphere
+atmos = jl.AGNI.atmosphere.Atmos_t()
+jl.AGNI.atmosphere.setup_b(atmos, ...)   # <--- complete function arguments as per docstring in `AGNI.atmosphere.setup!()`
+
+# Allocate atmosphere
+jl.AGNI.atmosphere.allocate_b(atmos, STAR_SPECTRUM_FILE)   # <-- provide path to spectrum
+
+# Solve T(p)
+jl.AGNI.solver.solve_energy_b(atmos)
+
+# Write results to a file
+jl.AGNI.save.write_ncdf(atmos, "out.nc")
+```
