+++ res/config/default.toml
@@ -22,8 +22,8 @@ title = "Default"                       # Name for this configuration file
     wind_speed      = 2.0               # Effective wind speed for sensible heat [m s-1].
 
 [files]
-    input_sf        = "res/spectral_files/Frostflow/256/Frostflow.sf"   # Path to SOCRATES spectral file.
-    input_star      = "res/stellar_spectra/trappist-1.txt"              # Path to stellar spectrum.
+    input_sf        = "res/spectral_files/Dayspring/48/Dayspring.sf"   # Path to SOCRATES spectral file.
+    input_star      = "res/stellar_spectra/sun.txt"              # Path to stellar spectrum.
     output_dir      = "out/"                                            # Path to output directory.
 
 [composition]
