### 2025-04-22 7ede878
Clarify solution types in docs
