### 2025-04-22 f1b1cd6
Clarify solution types in docs
