+++ .github/workflows/documentation.yml
@@ -22,13 +22,19 @@ jobs:
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v4
+
       - uses: julia-actions/setup-julia@v2
         with:
           version: '1.10'
       - uses: julia-actions/cache@v1
+
       - name: Install dependencies
         run: julia -e 'using Pkg; Pkg.add("Documenter"); Pkg.instantiate()'
+
+      - name: Install AGNI
+        run: julia --project="./" -e 'using Pkg; Pkg.add(path="./")'
+
       - name: Build and deploy
         env:
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }} # If authenticating with GitHub Actions token
-        run: julia docs/make.jl
+        run: julia --project="./" docs/make.jl
