+++ .github/workflows/documentation.yml
@@ -1,5 +1,4 @@
 # This job updates the documentation and pushes it to GitHub pages
-# This occurs when a PR to main is merged
 
 name: Docs
 
@@ -35,13 +34,43 @@ jobs:
           version: '1.10'
       - uses: julia-actions/cache@v1
 
-      - name: Install dependencies
+      - name: Install Documenter
         run: julia -e 'using Pkg; Pkg.add("Documenter"); Pkg.instantiate()'
 
-      - name: Install AGNI
-        run: julia -e 'using Pkg; Pkg.develop(path=pwd())'
+      # Setup SOCRATES
+      - name: Get SOCRATES
+        uses: actions/checkout@v4
+        with:
+          repository: 'nichollsh/SOCRATES'
+          path: 'SOCRATES'
 
-      - name: Build and deploy
+      - uses: actions/cache@v4
+        id: cache-socrates
+        with:
+          path: |
+            SOCRATES/bin
+            SOCRATES/sbin
+            SOCRATES/set_rad_env
+          key: socrates-${{ hashFiles('SOCRATES/version') }}
+
+      - name: Build SOCRATES
+        if: steps.cache-socrates.outputs.cache-hit != 'true'
+        run: |
+          export LD_LIBRARY_PATH=""
+          export RAD_DIR="/home/runner/work/AGNI/AGNI/SOCRATES"
+          cd $RAD_DIR
+          ./configure
+          ./build_code
+          cd ..
+
+      - name: Build AGNI
+        run: |
+          export RAD_DIR="/home/runner/work/AGNI/AGNI/SOCRATES"
+          export RAD_BIN="$RAD_DIR/bin/"
+          export LD_LIBRARY_PATH=""
+          julia --project="./" -e 'using Pkg; Pkg.build()'
+
+      - name: Build docs and deploy
         env:
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }} # If authenticating with GitHub Actions token
-        run: julia docs/make.jl
+        run: julia --project="./" docs/make.jl
