### 2025-04-22 170be6b
Clarify fastchem coupling in docs
