### 2025-04-22 6e73904
Clarify fastchem coupling in docs
