### 2025-04-23 1227afa
Enforce OpenSSL version 3.0
