### 2025-04-23 2829436
Strictly enforce OpenSSL 3.0
