+++ Project.toml
@@ -27,6 +27,6 @@ Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
 TOML = "fa267f1f-6049-4f14-aa54-33bafae1ed76"
 
 [compat]
-OpenSSL_jll = "3.0.16"
+OpenSSL_jll = "= 3.0.16"
 SpecialFunctions = "2.5.0"
 julia = ">= 1.10"
