### 2025-04-23 b33b755
Update get_data with generic spectral file lookup. Update docs
