### 2025-04-23 171cc4d
better error message when missing spfile
