+++ res/config/cloud.toml
@@ -14,10 +14,8 @@ title = "Cloud radiative properties test"
     flux_int        = 0.0
 
 [files]
-    # input_sf        = "res/spectral_files/Frostflow/256/Frostflow.sf"
-    # input_star      = "res/stellar_spectra/sun.txt"
-    input_sf  = "out/runtime.sf"
-    input_star = ""
+    input_sf        = "res/spectral_files/Frostflow/256/Frostflow.sf"
+    input_star      = "res/stellar_spectra/sun.txt"
     output_dir      = "out/"
 
 [composition]
@@ -36,7 +34,7 @@ title = "Cloud radiative properties test"
     num_levels      = 100
     continua        = true
     rayleigh        = false
-    cloud           = false
+    cloud           = true
     aerosol         = false
     overlap_method  = "ro"
     real_gas        = false
