+++ docs/src/setup.md
@@ -43,11 +43,11 @@ bash src/get_agni.sh
 ```
 
 ## Using the code
-See [Running the model](@ref) for information on using the code.
+See [Using the model](@ref) for information on using the code.
 See [Troubleshooting](@ref) for troubleshooting advice.
 
 ## Coupling with FastChem
-This can be enabled using the configuration file parameter `composition.chem_type`. Of
+This can be enabled using the configuration file parameter `composition.chemistry`. Of
 course, it is first necessary to set up FastChem, which can be done by running:
 ```bash
 bash src/get_fastchem.sh
@@ -55,4 +55,4 @@ bash src/get_fastchem.sh
 You **must** then set the `FC_DIR` environment variable to the location of the FastChem
 installation folder. Ideally you should also add this variable to your bashrc file.
 
-There is information 
+There is information
