### 2025-04-24 c9540c3
Update configs (remove include_all and flag_aerosol). Update docs to explain config parameters. Cast types
