+++ docs/src/usage.md
@@ -1,4 +1,4 @@
-# Running the model
+# Using the model
 First, follow the [Getting started](@ref) instructions. Only read on once you
 have confirmed that the code is working.
 
@@ -71,16 +71,96 @@ There are no 'default values'. Not all parameters are required in all cases,
 but the model will return an error naming any parameters which are both
 necessary and absent.
 
-Broadly, the configuration files are broken up into four sections:
-* `[planet]` -  general characteristics of the planet
-* `[files]` - input/output files and other paths
-* `[composition]` - atmospheric composition and chemistry
-* `[execution]` - what the model should do
-* `[plots]` - which plots should be produced
+The configuration files are broken up into four sections (or "tables"), each containing a
+number of parameters. These tables of parameters are described below.
 
-Some parameters:
-* `files.input_sf` is the file path to the "spectral file" containing opacity data. Several spectral files are packged with AGNI, but you can find more online via the [Open Science Framework](https://osf.io/vehxg/).
+### `[planet]`
+General properties of the planet.
+| Parameter          | Description   |
+| ------------------ | ------------- |
+| `tmp_surf        ` | Temperature of the surface [kelvin].  |
+| `instellation    ` | Stellar flux at planet's orbital distance [W m-2]. |
+| `albedo_b        ` | A pseudo bond-albedo which downscales the stellar flux by 1-albedo_b. |
+| `s0_fact         ` | Stellar flux scale factor which accounts for planetary rotation (c.f. Cronin+13). |
+| `zenith_angle    ` | Characteristic zenith angle for incoming stellar radiation [degrees]. |
+| `surface_material` | Surface material (can be "greybody" or path to data file in `res/`). |
+| `albedo_s        ` | Spectrally-grey surface albedo used when `material=greybody`. Optional otherwise. |
+| `radius          ` | Radius of the planet's surface [m]. |
+| `gravity         ` | Gravitational acceleration at the surface [m s-2]. Incompatible with `mass` option. |
+| `mass            ` | Total mass of material below the atmosphere [kg]. Incompatible with `gravity` option. |
+| `skin_d          ` | Thickness of the conductive boundary layer [m]. Used when `solution_type=2`. |
+| `skin_k          ` | Conductivity of the conductive boundary layer [W m-1 K-1]. Used when `solution_type=2`. |
+| `tmp_magma       ` | Temperature of the topmost layer of the planet's mantle [K]. Used when `solution_type=2`. |
+| `flux_int        ` | Internal flux [W m-2] to be solved-for when `solution_type=3`. |
+| `turb_coeff      ` | Turbulent exchange coefficient for sensible heat transport. |
+| `wind_speed      ` | Effective wind speed for sensible heat transport [m s-1]. |
 
+
+### `[files]`
+Input/output files and other paths.
+| Parameter          | Description   |
+| ------------------ | ------------- |
+| `input_sf       `  | Path to the desired spectral file ending in `.sf`, in `res/spectral_files/`. |
+| `input_star     `  | Path to stellar spectrum file. Spectrum assumed to be inside spectral file if this is left blank. |
+| `output_dir     `  | Path to the output directory. |
+
+
+### `[composition]`
+Atmospheric composition and chemistry.
+| Parameter         | Description   |
+| ----------------- | ------------- |
+| `p_top         `  | Total top-of-atmosphere pressure [bar]. |
+| `p_dict        `  | Dictionary of gas partial surface pressures [bar]. Summed to obtain `p_surf`. |
+| `p_surf        `  | Total surface pressure [bar]. Incompatible with `p_dict`.|
+| `vmr_dict      `  | Gas volume mixing ratios (=mole fractions) at the surface. Must be set alongside `p_surf`. |
[…]
