### 2025-04-25 #7726 @nichollsh
@arfon I have a response prepared to the above review, as well as improvements to the code, paper, and documentation. Please advise on whether I am expected to respond to the reviewers individually/directly, or respond to you once a decision has been made.
