### 2025-04-25 78d4e0e
Light vs dark logos
