### 2025-04-25 c222650
Fix phrasing in paper
