+++ docs/paper/paper.bib
@@ -373,6 +373,14 @@ archivePrefix = {arXiv},
         title = "Self-limited tidal heating and prolonged magma oceans in the L 98-59 system",
       journal = {In review at MNRAS},
          year = 2025,
+         month = {Feb}
+}
+
+@article{nicholls_escape_2025,
+       author = {Harrison Nicholls and Tim Lichtenberg and Richard Chatterjee and Emma Postolec and Raymond Pierrehumbert},
+        title = "A sub-Neptunian origin of the super-Earth L 98-59 d",
+      journal = {In preparation},
+         year = 2025,
          month = {June}
 }
 
