### 2025-04-25 943c659
Fix table formatting in docs
