### 2025-04-25 bda4a3a
Add logo for docs and favicon
