### 2025-04-25 4c56694
Fix condense.toml and update docs with info on structure
