### 2025-04-25 c43c43b
Fix condense.toml and update docs with info on structure
