+++ res/config/condense.toml
@@ -35,7 +35,7 @@ title = "Condensation test"
     verbosity       = 1
     max_steps       = 300
     max_runtime     = 400
-    num_levels      = 35
+    num_levels      = 50
     continua        = true
     rayleigh        = false
     cloud           = false
@@ -45,14 +45,14 @@ title = "Condensation test"
     sensible_heat   = true
     latent_heat     = true
     convection      = true
-    rainout         = true
+    rainout         = false
     solution_type   = 3
     solver          = "newton"
-    dx_max          = 200.0
-    initial_state   = ["loglin","1000"]
-    linesearch      = 2
+    dx_max          = 300.0
+    initial_state   = ["loglin","1100"]
+    linesearch      = 1
     easy_start      = true
-    converge_atol   = 0.1
+    converge_atol   = 1.0
     converge_rtol   = 1.0e-3
 
 
