### 2025-04-25 #7726 @nichollsh
Hi @arjunsavel,

Thanks again for the positive review. I appreciate your constructive feedback on my model, its documentation, and the manuscript.

I've broken down my response into the same headings as your review. Most of these changes were made as part of [this PR](https://github.com/nichollsh/AGNI/pull/97).

## Paper

* I have added a discussion of how my model compares to those used by Piette+2023 and Zilinskas+2023. These both solve for radiative-convective equilibrium, but the former is proprietary and the latter only considers rock-vapour atmospheres.
* Corrected grammar and typos pointed out by the reviewer.
* While a including a figure in the paper would indeed be useful, I have elected to leave one out at the present. This is because the documentation includes examples already, with an animation, and the paper is additionally constrained by the journal's formatting requirements.

## Documentation

* The theoretical aspect of coupling with PROTEUS is described and demonstrated within Nicholls+[2025](https://academic.oup.com/mnras/article/536/3/2957/7926963). I have added information in the docs which describes how this is done in practice with `juliacall`, and directed the reader to the location in the PROTEUS repository where this is implemented.
* Corrected grammar and typos.
* Added information into the docs as to when each solution type could be used.
* Clarified implementation of convection with MLT.
* "Why must the initial [guess of the T(p) solution] not be significantly cooler [than the true solution]?" I believe that this is primarily a result of radiative diffusion scaling as temperature cubed. This means that the model has greater sensitivity (and larger numerical gradients) at higher temperatures, which can help the solver to determine the correct 'direction' at each iteration.
* Defined TKE in the docs and in the paper ("turbulent kinetic energy").
* The jacobian matrix expresses how each element of the residual $r$ vector (i.e. flux lost each each level) changes with the temperature $x$ at each level ($J_{ij} = dr_i / dx_j$). The energy flux calculation (including radiative transfer, MLT, etc.) evaluates the $r$ vector for a given $x$ vector. This derivative is evaluated with with finite-differences by looping over each level $j$ and perturbing the temperatures individually, thereby yielding a whole row of $J$. This row represents the sensitivity of the energy fluxes at all layers to the relative change in temperature at a given layer $j$. We therefore need $N$ evaluations to fill the $N \times N$ matrix completely.
* Clarified purpose of $C_a$ and $C_r$
* I have added functionality documentation [here](https://nichollsh.github.io/AGNI/dev/manual/#Code-reference), based on the autodoc feature of Documenter.jl
* Indeed, the atmosphere is assumed to have the abundances provided by the user when chemistry is turned off. This can be constant, or from a CSV file. Updated docs to clarify.
* Added a section into the [usage](https://nichollsh.github.io/AGNI/dev/usage/#Configuration) page of the docs which describes the purpose of each parameter. This includes units where appropriate, and information on how the parameters relate to each other.

## Notes from running
* "... realpath: bash: No such file or directory". Unfortunately I am not able to reproduce this behaviour.
* Made `get_socrates.sh` more robust by using HTTPS to clone the repo rather than SSH.
* Updated cloud configuration file so that it does not crash.
* Updated condensation configuration so that it works again.
* I have made the `get_data.sh` script more flexible by allowing any spectral file to be downloaded from OSF. There's also a table included in the "basic" dataset which describes the available spectral files, and how they compare to each other. Added a [new section](https://nichollsh.github.io/AGNI/dev/usage/#Input-data-files) in the docs explaining this.
* "Spectral files contain no aerosol data". I've clarified the content of the spectral files in the documentation. They indeed contain cloud droplet data, rayleigh scattering cross-sections, gas opacities, and continuum opacities. However, I am not able to reproduce this error.
* I have added a section in the docs explaining what the gas flags mean. When `NO_OPACITY` is printed, it indeed means that the gas does not contribute radiatively but can still contribute to the thermodynamics (i.e. it has some density, heat capacity, etc.). There are similar flags such as `NO_THERMO` which are now clarified in the docs.

Thank you for considering my paper for JOSS.
I am looking forward to your response @arjunsavel, and to your review @MartianColonist.

-Harrison
