### 2025-04-29 2f97942
Add transparent solution mode
