+++ src/AGNI.jl
@@ -255,6 +255,18 @@ module AGNI
         p_surf::Float64                 = 0.0
         mf_dict::Dict{String, Float64}  = Dict{String, Float64}()
         mf_path::String                 = ""
+        transparent::Bool               = false
+
+        # transparent atmosphere
+        if haskey(cfg["composition"], "transparent")
+            transparent = Bool(cfg["composition"]["transparent"])
+
+            # set dummy values
+            mf_dict = Dict("N2"=>1.0,"H2O"=>0.0,"CO2"=>0.0,"S2"=>0.0)
+            p_surf  = 1.0
+        end
+
+        # composition set by VMRs
         if haskey(cfg["composition"],"p_surf")
             # set composition using VMRs + Psurf
             if haskey(cfg["composition"],"vmr_dict")
@@ -269,6 +281,7 @@ module AGNI
             end
             p_surf = cfg["composition"]["p_surf"]
 
+        # composition set by partial pressures
         elseif haskey(cfg["composition"], "p_dict")
             # set composition from partial pressures (converted to mixing ratios)
             pp_dict::Dict{String, Float64} = cfg["composition"]["p_dict"]
@@ -279,17 +292,25 @@ module AGNI
                 mf_dict[k] = pp_dict[k]/p_surf
             end
 
-        else
-            @error "Misconfiguration: must provide either p_dict or p_surf+VMRs"
+        # most provide either VMR or partial pressures, if not transparent
+        elseif !transparent
+            @error "Must provide either p_dict or p_surf+VMRs in config"
+            @error "    Neglect these keys only when transparent=true"
             return false
         end
+
         if chem_type in [1,2,3]
             if length(condensates)>0
-                @error "Misconfiguration: chemistry coupling is incompatible with condensation"
+                @error "Chemistry coupling is incompatible with condensation"
                 return false
-            else
-                mkdir(dir_fastchem)
             end
+
+            if transparent
+                @error "Chemistry couplng is incompatible with transparent atmosphere"
+                return false
+            end
+
+            mkdir(dir_fastchem)
         end
 
         #    solver stuff
@@ -298,14 +319,8 @@ module AGNI
         incl_sens::Bool        =        cfg["execution"]["sensible_heat"]
         incl_latent::Bool      =        cfg["execution"]["latent_heat"]
         sol_type::Int          =        cfg["execution"]["solution_type"]
-        dx_max::Float64        =        cfg["execution"]["dx_max"]
-        linesearch::Int        =        cfg["execution"]["linesearch"]
-        easy_start::Bool       =        cfg["execution"]["easy_start"]
         conv_atol::Float64     =        cfg["execution"]["converge_atol"]
         conv_rtol::Float64     =        cfg["execution"]["converge_rtol"]
-        max_steps::Int         =        cfg["execution"]["max_steps"]
-        max_runtime::Float64   =        cfg["execution"]["max_runtime"]
-        rainout::Bool          =        cfg["execution"]["rainout"]
 
         #    plotting stuff
         plt_tmp::Bool          = cfg["plots"]["temperature"]
@@ -376,6 +391,11 @@ module AGNI
         # Allocate atmosphere
         return_success = atmosphere.allocate!(atmos,star_file) || return false
 
+        # Set transparent
+        if transparent
+            atmosphere.make_transparent!(atmos)
+        end
+
         # Set T(p) by looping over requests
         # Each request may be a command, or an argument following a command
         setpt.request!(atmos, cfg["execution"]["initial_state"]) || return false
@@ -403,7 +423,6 @@ module AGNI
         end
 
[…]
