### 2025-04-29 51fafc0
Fix config. Add docs explaining transparent solver. Pass params to transparent solver. Version 1.5.0
