+++ src/AGNI.jl
@@ -253,66 +253,76 @@ module AGNI
         end
 
         #    composition stuff
-        condensates::Array{String,1}    = cfg["composition"]["condensates"]
-        chem_type::Int                  = cfg["composition"]["chemistry"]
+        condensates::Array{String,1}    = String[]
+        chem_type::Int                  = 0
         p_surf::Float64                 = 0.0
+        p_top::Float64                  = 0.0
         mf_dict::Dict{String, Float64}  = Dict{String, Float64}()
         mf_path::String                 = ""
         transparent::Bool               = false
+        real_gas::Bool                  = false
 
-        # transparent atmosphere
+        # transparent atmosphere?
         if haskey(cfg["composition"], "transparent")
             transparent = Bool(cfg["composition"]["transparent"])
+        end
+        if transparent
+            @info "Transparent atmosphere requested"
 
             # set dummy values
-            mf_dict = Dict("N2"=>1.0,"H2O"=>0.0,"CO2"=>0.0,"S2"=>0.0)
-            p_surf  = 1.0
-        end
+            mf_dict  = Dict("N2"=>1.0,"H2O"=>0.0,"CO2"=>0.0,"S2"=>0.0)
+            p_surf   = 1e-3
+            p_top    = 1e-5
+            real_gas = false
 
-        # composition set by VMRs
-        if haskey(cfg["composition"],"p_surf")
-            # set composition using VMRs + Psurf
-            if haskey(cfg["composition"],"vmr_dict")
-                # from dict in cfg file
-                mf_dict = cfg["composition"]["vmr_dict"]
-            elseif haskey(cfg["composition"], "vmr_path")
-                # from csv file to be read-in
-                mf_path = cfg["composition"]["vmr_file"]
+
+        # not transparent
+        else
+            p_top = Float64(cfg["composition"]["p_top"])
+            real_gas = Bool(cfg["execution"]["real_gas"])
+            chem_type = Int(cfg["composition"]["chemistry"])
+            condensates = cfg["composition"]["condensates"]
+
+            # composition set by VMRs
+            if haskey(cfg["composition"],"p_surf")
+                # set composition using VMRs + Psurf
+                if haskey(cfg["composition"],"vmr_dict")
+                    # from dict in cfg file
+                    mf_dict = cfg["composition"]["vmr_dict"]
+                elseif haskey(cfg["composition"], "vmr_path")
+                    # from csv file to be read-in
+                    mf_path = cfg["composition"]["vmr_file"]
+                else
+                    @error "Misconfiguration: if providing p_surf, must also provide VMRs"
+                    return false
+                end
+                p_surf = cfg["composition"]["p_surf"]
+
+            # composition set by partial pressures
+            elseif haskey(cfg["composition"], "p_dict")
+                # set composition from partial pressures (converted to mixing ratios)
+                pp_dict::Dict{String, Float64} = cfg["composition"]["p_dict"]
+                for k in keys(pp_dict)
+                    p_surf += pp_dict[k]
+                end
+                for k in keys(pp_dict)
+                    mf_dict[k] = pp_dict[k]/p_surf
+                end
+
+            # most provide either VMR or partial pressures, if not transparent
             else
-                @error "Misconfiguration: if providing p_surf, must also provide VMRs"
+                @error "Must provide either p_dict or p_surf+VMRs in config"
+                @error "    Neglect these keys only when transparent=true"
                 return false
             end
-            p_surf = cfg["composition"]["p_surf"]
-
-        # composition set by partial pressures
-        elseif haskey(cfg["composition"], "p_dict")
-            # set composition from partial pressures (converted to mixing ratios)
[…]
