### 2025-04-29 12000f4
Fix config. Add docs explaining transparent solver. Pass params to transparent solver. Version 1.5.0
