### 2025-04-29 9e727fc
Catch error from OSF
