### 2025-04-29 132f34a
Bugfix in GS search method
