### 2025-04-29 5533143
Bugfix in GS search method
