### 2025-05-05 #7726 @nichollsh
Hi all, @arfon @MartianColonist @arjunsavel,

We are coming up on 8 months since I first submitted this paper for review in JOSS.

Is there anything I can do to help speed things up here? Please let me know if you're having any problems using the code or accessing the draft, and I will be very happy to help you.

-Harrison
