### 2025-05-05 #7726 @MartianColonist
Hi @nichollsh

Thank you for your patience during the review process (I'm in the middle of moving countries right now, so sorry for the slow response time).

This first part of my review offers feedback on the paper. The paper overall is clear, well-written, and AGNI will make a great addition to JOSS. My feedback and suggestions below will help improve the completeness of the paper before publication in JOSS.

**Feedback on the software paper**

_General comments_:

- There are a few missing sections commonly seen in JOSS papers:

1. **Similar Tools**: while there is a section 'Comparison with other codes', the [JOSS docs](https://joss.readthedocs.io/en/latest/paper.html) require "A list of key references, including to other software addressing related needs"). You can simply add a new section with the names, citations, and links to the code repositories of other similar open source codes.
2. **Documentation**: a single-sentence section pointing the user to the AGNI documentation. This is already stated on Line 25, but should be moved to a stand-alone section.
3. **Future Developments** (optional): while not necessary, it would be good to mention any planned future developments / new applications of AGNI after the public release corresponding to the JOSS paper.

- I recommend including a Figure in the paper demonstrating a typical application of AGNI. It is common in JOSS papers to include a Figure showing a flowchart or schematic diagram that illustrates what the code is used for. For example, an artist concept of a lava world with an arrow pointing to the AGNI logo, then an arrow pointing from the logo to a simulated spectrum (such as one of the plots in the AGNI documentation). This would be a nice visual advertisement for AGNI and would enhance the paper.

- The Statement of Need section is currently missing some important details about how AGNI compares to other publicly available codes. While the section ‘Comparison with other codes’ outlines features of AGNI and states that AGNI uses a Newton-Raphson method for faster convergence than some other codes, there is not a direct comparison in terms of how the applications of AGNI relate to other previous work. I suggest adding a few sentences to the Statement of Need / Comparison with Other Codes stating:

1. Which other lava planet modelling codes are open source?
2. What is the ‘gap’ that AGNI fills (i.e. are there no open source codes similar to AGNI? Are the existing open source codes too physics-limited? Are the existing open source codes too slow?).
3. Who is the intended user / target audience for AGNI?

_Line-by-line specific feedback_:

- Line 2: Remove the full stop at the end of the title.
- Line 14: Add reference to the “well established FORTRAN code”.
- Line 37: Typo: “contrained” → “constrained”.
- Line 38: Typo: “exoplanet TRAPPIST-1 b” → “the exoplanet TRAPPIST b”.
- Line 44: Typo: “with radiative transfer with an interior evolution model” → “with radiative transfer and an interior evolution model”.
- Line 78: “operate with sufficient speed that it may participate in a wide parameter space” – how much compute time does a single model evaluation typically take?

Great work @nichollsh on the paper!

I will now go through the code documentation and verify that AGNI installs and runs on my Linux machine. I will provide further feedback on the code in a separate comment and/or via issues raised in the AGNI GitHub repository.
