+++ docs/paper/paper.bib
@@ -18,6 +18,23 @@
   institution={Met Office, Exeter, UK}
 }
 
+@article{gandhi_genesis_2017,
+    author = {Gandhi, Siddharth and Madhusudhan, Nikku},
+    title = {genesis: new self-consistent models of exoplanetary spectra},
+    journal = {Monthly Notices of the Royal Astronomical Society},
+    volume = {472},
+    number = {2},
+    pages = {2334-2355},
+    year = {2017},
+    month = {06},
+    abstract = {We are entering the era of high-precision and high-resolution spectroscopy of exoplanets. Such observations herald the need for robust self-consistent spectral models of exoplanetary atmospheres to investigate intricate atmospheric processes and to make observable predictions. Spectral models of plane-parallel exoplanetary atmospheres exist, mostly adapted from other astrophysical applications, with different levels of sophistication and accuracy. There is a growing need for a new generation of models custom-built for exoplanets and incorporating state-of-the-art numerical methods and opacities. The present work is a step in this direction. Here we introduce genesis, a plane-parallel, self-consistent, line-by-line exoplanetary atmospheric modelling code that includes (a) formal solution of radiative transfer using the Feautrier method, (b) radiative–convective equilibrium with temperature correction based on the Rybicki linearization scheme, (c) latest absorption cross-sections, and (d) internal flux and external irradiation, under the assumptions of hydrostatic equilibrium, local thermodynamic equilibrium and thermochemical equilibrium. We demonstrate the code here with cloud-free models of giant exoplanetary atmospheres over a range of equilibrium temperatures, metallicities, C/O ratios and spanning non-irradiated and irradiated planets, with and without thermal inversions. We provide the community with theoretical emergent spectra and pressure–temperature profiles over this range, along with those for several known hot Jupiters. The code can generate self-consistent spectra at high resolution and has the potential to be integrated into general circulation and non-equilibrium chemistry models as it is optimized for efficiency and convergence. genesis paves the way for high-fidelity remote sensing of exoplanetary atmospheres at high resolution with current and upcoming observations.},
+    issn = {0035-8711},
+    doi = {10.1093/mnras/stx1601},
+    url = {https://doi.org/10.1093/mnras/stx1601},
+    eprint = {https://academic.oup.com/mnras/article-pdf/472/2/2334/19942998/stx1601.pdf},
+}
+
+
 @article{malik_helios_2017,
 	title = {{HELIOS}: {AN} {OPEN}-{SOURCE}, {GPU}-{ACCELERATED} {RADIATIVE} {TRANSFER} {CODE} {FOR} {SELF}-{CONSISTENT} {EXOPLANETARY} {ATMOSPHERES}},
 	volume = {153},
@@ -32,6 +49,20 @@
 	month = jan,
 	year = {2017},
 	pages = {56},
+}
+
+@article{mukherjee_picaso_2023,
+doi = {10.3847/1538-4357/ac9f48},
+url = {https://dx.doi.org/10.3847/1538-4357/ac9f48},
+year = {2023},
+month = {jan},
+publisher = {The American Astronomical Society},
+volume = {942},
+number = {2},
+pages = {71},
+author = {Mukherjee, Sagnick and Batalha, Natasha E. and Fortney, Jonathan J. and Marley, Mark S.},
+title = {PICASO 3.0: A One-dimensional Climate Model for Giant Planets and Brown Dwarfs},
+journal = {The Astrophysical Journal},
 }
 
 
@@ -56,6 +87,30 @@
 	year = {2014},
 	pages = {A59},
 }
+
+@Article{wordsworth_coupled_2021,
+author="Wordsworth, Robin
[…]
