+++ docs/paper/paper.md
@@ -35,7 +35,7 @@ It is important that we are able to accurately model the atmospheres of (exo)pla
 
 AGNI[^1] is a Julia program designed to solve for the temperature and radiation environment within the atmospheres of rocky (exo)planets. It leverages a well established FORTRAN code to calculate radiative fluxes from a given atmospheric temperature structure and composition, which -- alongside representations of convection and other processes -- enables an energy-conserving numerical solution for the atmospheric conditions. In contrast to most other numerical atmosphere codes, AGNI uses an Newton-Raphson optimisation method to obtain this solution, which enables improved performance and scalability. AGNI was specifically developed for use alongside planetary interior models within a coupled framework, although it can also be easily applied to scientific problems as a standalone code.
 
-AGNI can be interacted with as a library; it is used in this way within the [Jupyter notebook tutorials](https://github.com/nichollsh/AGNI/tree/main/tutorials) in the repository. It can also be used as an executable program, where it reads TOML configuration files from the disk and outputs figures and NetCDF data to a specified directory. The documentation can be read online [here](https://nichollsh.github.io/AGNI/).
+AGNI can be interacted with as a library; it is used in this way within the [Jupyter notebook tutorials](https://github.com/nichollsh/AGNI/tree/main/tutorials) in the repository. It can also be used as an executable program, where it reads TOML configuration files from the disk and outputs figures and NetCDF data to a specified directory.
 
 [^1]: AGNI can be found on GitHub [here](https://github.com/nichollsh/AGNI).
 
@@ -45,29 +45,40 @@ It is thought that all rocky planets go through a "magma ocean" stage, where the
 
 Several theoretical studies have modelled the atmospheres and evolution of these young planets, but all have made several simplifying assumptions. @lichtenberg_vertically_2021 coupled a simple atmosphere climate model with radiative transfer with an interior evolution model to simulate magma ocean evolution, but they did not account for the possibility of convective stability in their atmospheres. @krissansen_was_2021 and @krissansen_erosion_2024 made similar assumptions. @selsis_cool_2023 used a pure-steam radiative-convective model to investigate convective stability, finding that it is likely to occur, but did not extend their work to coupled scenarios which explore the secular evolution of these planets or gas mixtures; it is unlikely that these atmospheres are exclusively composed of steam [@lichtenberg_vertically_2021]. @piette_rocky_2023 similarly explored potential atmospheres on observable lava planets, but did not consider the physics of atmosphere-interior coupling, and instead chose semi-arbitrary gas compositions. @zilinskas_observability_2023 used the HELIOS atmosphere model to model potential rock-vapour envelopes of sub-Neptune and super-Earth exoplanets, finding that the opacity of various gaseous species (notably SiO) plays a key role in determining the structure and observable properties of exoplanet atmospheres, although they neglect volatile species and apply the ideal gas equations of state. The demand for realistic atmosphere modelling in the context of secular magma ocean evolution is apparent.
 
[…]
