### 2025-05-06 5105218
Remove table
