+++ docs/paper/paper.bib
@@ -76,8 +76,8 @@ journal = {The Astrophysical Journal},
 
 
 @article{amundsen_treatment_2017,
-	author = {{Amundsen, David S.} and {Tremblin, Pascal} and {Manners, James} and {Baraffe, Isabelle} and {Mayne, Nathan J.}},
-	title = {Treatment of overlapping gaseous absorption   with the correlated-k method in hot Jupiter   and brown dwarf atmosphere models},
+	author = {Amundsen, David S. and Tremblin, Pascal and Manners, James and Baraffe, Isabelle and Mayne, Nathan J.},
+	title = {Treatment of overlapping gaseous absorption with the correlated-k method in hot Jupiter and brown dwarf atmosphere models},
 	DOI= "10.1051/0004-6361/201629322",
 	url= "https://doi.org/10.1051/0004-6361/201629322",
 	journal = {A&A},
@@ -148,22 +148,14 @@ URL = {https://gmd.copernicus.org/articles/16/5601/2023/},
 DOI = {10.5194/gmd-16-5601-2023}
 }
 
-@ARTICLE{zieba_obs_2023,
-       author = {{Zieba}, Sebastian and {Kreidberg}, Laura and {Ducrot}, Elsa and {Gillon}, Micha{\"e}l and {Morley}, Caroline and {Schaefer}, Laura and {Tamburo}, Patrick and {Koll}, Daniel D.~B. and {Lyu}, Xintong and {Acu{\~n}a}, Lorena and {Agol}, Eric and {Iyer}, Aishwarya R. and {Hu}, Renyu and {Lincowski}, Andrew P. and {Meadows}, Victoria S. and {Selsis}, Franck and {Bolmont}, Emeline and {Mandell}, Avi M. and {Suissa}, Gabrielle},
-        title = "{No thick carbon dioxide atmosphere on the rocky exoplanet TRAPPIST-1 c}",
-      journal = {Nature},
-     keywords = {Astrophysics - Earth and Planetary Astrophysics},
-         year = 2023,
-        month = aug,
-       volume = {620},
-       number = {7975},
-        pages = {746-749},
-          doi = {10.1038/s41586-023-06232-z},
-archivePrefix = {arXiv},
-       eprint = {2306.10150},
- primaryClass = {astro-ph.EP},
-       adsurl = {https://ui.adsabs.harvard.edu/abs/2023Natur.620..746Z},
-      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+@misc{belloarufe_so2_2025,
+      title={Evidence for a volcanic atmosphere on the sub-Earth L98-59b},
+      author={Aaron Bello-Arufe and Mario Damiano and Katherine A. Bennett and Renyu Hu and Luis Welbanks and Ryan J. MacDonald and Darryl Z. Seligman and David K. Sing and Armen Tokadjian and Apurva Oza and Jeehyun Yang},
+      year={2025},
+      eprint={2501.18680},
+      archivePrefix={arXiv},
+      primaryClass={astro-ph.EP},
+      url={https://arxiv.org/abs/2501.18680},
 }
 
 @ARTICLE{hu_cancri_2024,
@@ -415,6 +407,24 @@ archivePrefix = {arXiv},
        adsurl = {https://ui.adsabs.harvard.edu/abs/2025ApJ...978L..40H},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
+
+@article{nicholls_redox_2024,
+author = {Nicholls, Harrison and Lichtenberg, Tim and Bower, Dan J. and Pierrehumbert, Raymond},
+title = {Magma Ocean Evolution at Arbitrary Redox State},
[…]
