+++ docs/paper/paper.md
@@ -31,38 +31,36 @@ bibliography: paper.bib
 
 # Summary
 
-It is important that we are able to accurately model the atmospheres of (exo)planets. This is because their atmospheres play a crucial role in setting the environment and conditions on the planet, including how the planet evolves over astronomical timescales. Additionally, it is primarily by observation of their atmospheres that we are able to characterise exoplanets. There is demand for accurate atmosphere models in the context of lava worlds: planets with permanent or fleeting magma oceans.
+It is important that we are able to accurately model the atmospheres of (exo)planets. This is because atmospheres play a central role in setting a planet's thermochemical environment at a given point in time, and also in regulating how it evolves over geological timescales. Additionally, it is primarily by observation of their atmospheres that we are able to characterise exoplanets. There is particular demand for accurate models in the context of so-called lava worlds: planets with molten interiors (or 'magma oceans').
 
-AGNI[^1] is a Julia program designed to solve for the temperature and radiation environment within the atmospheres of rocky (exo)planets. It leverages a well established FORTRAN code [@edwards_studies_1996; @sergeev_socrates_2023] to calculate radiative fluxes from a given atmospheric temperature structure and composition, which -- alongside representations of convection and other processes -- enables an energy-conserving numerical solution for the atmospheric conditions. In contrast to most other numerical atmosphere codes, AGNI uses an Newton-Raphson optimisation method to obtain this solution, which enables improved performance and scalability. AGNI was specifically developed for use alongside planetary interior models within a coupled framework, although it can also be easily applied to scientific problems as a standalone code.
-
-AGNI can be interacted with as a library; it is used in this way within the [Jupyter notebook tutorials](https://github.com/nichollsh/AGNI/tree/main/tutorials) in the repository. It can also be used as an executable program, where it reads TOML configuration files from the disk and outputs figures and NetCDF data to a specified directory.
+AGNI is a Julia program designed to solve for the temperature and radiation environment within the atmospheres of rocky (exo)planets. It leverages a well established FORTRAN code [@edwards_studies_1996; @sergeev_socrates_2023] to calculate radiative fluxes from a given atmospheric temperature structure and composition, which -- alongside representations of convection and other processes -- enables an energy-conserving numerical solution for the atmospheric conditions. In contrast to most other numerical atmosphere models, AGNI uses a Newton-Raphson optimisation method to obtain its solution, which enables improved performance and scalability. Our model was specifically developed for use alongside planetary interior models within a coupled simulation framework. However, it can also be applied to scientific problems standalone when used as an executable program; it reads TOML configuration files and outputs figures and NetCDF datasets. AGNI can also function as a software library; it is used in this sense within the [Jupyter notebook tutorials](https://github.com/nichollsh/AGNI/tree/main/tutorials) of our GitHub repository[^1].
 
 [^1]: AGNI can be found on GitHub [here](https://github.com/nichollsh/AGNI).
 
 # Statement of need
 
[…]
