### 2025-05-07 551e9b3
Update citation for tides paper
