+++ docs/paper/paper.bib
@@ -442,12 +442,14 @@ year = {2024}
     eprint = {https://academic.oup.com/mnras/article-pdf/536/3/2957/61218697/stae2772.pdf},
 }
 
-@article{nicholls_tidal_2025,
-       author = {Harrison Nicholls and Claire Guimond and Hamish Hay and Richard Chatterjee and Tim Lichtenberg and Raymond Pierrehumbert},
-        title = "Self-limited tidal heating and prolonged magma oceans in the L 98-59 system",
-      journal = {In review at MNRAS},
-         year = 2025,
-         month = {Feb}
+@misc{nicholls_tidal_2025,
+      title={Self-limited tidal heating and prolonged magma oceans in the L 98-59 system},
+      author={Harrison Nicholls and Claire Marie Guimond and Hamish C. F. C. Hay and Richard D. Chatterjee and Tim Lichtenberg and Raymond T. Pierrehumbert},
+      year={2025},
+      eprint={2505.03604},
+      archivePrefix={arXiv},
+      primaryClass={astro-ph.EP},
+      url={https://arxiv.org/abs/2505.03604},
 }
 
 @article{nicholls_escape_2025,
