### 2025-05-07 #7726 @nichollsh
Hi @MartianColonist,

Thank you for the constructive comments on my manuscript. I have updated it following your suggestions.

The most significant change is the inclusion of a figure (also suggested by @arjunsavel), which compares the two primary use cases for AGNI. This contextualises the model relative to the PROTEUS framework and compares this coupled application to using the code stand-alone.

I have also added a discussion which compares AGNI to other codes which have been applied to lava planet atmospheres in the literature. You can find this within the "Comparison with other codes" section.

There are some other minor changes - mostly updating citations and improving the grammar.

Thank you for taking the time to review AGNI and my paper.
