### 2025-05-08 539b908
Update docs and safety check on get_socrates
