### 2025-05-08 #7726 @arjunsavel
Hi @nichollsh — thanks for your thorough responses to my comments! They are more than satisfactory.

I have a couple more very minor paper comments:
- I appreciate the addition of the new figure. In the bottom-right cell, the text runs quite close to the edge — could this be fixed? Also, minor typo: "solves for atmosphere as requested" --> "solves for the atmospheric structure as requested"
- of a wide parameter space, --> of a wide parameter space.

Unrelated to review: I'll also +1 the future development of hazes and aerosols. Flexibility and user-friendliness on this front would be a great benefit, especially compared to other RCE codes.
