### 2025-05-08 2d3f3c6
Add check for required libraries
