### 2025-05-08 0d22b15
Footnotes
