+++ docs/paper/paper.bib
@@ -158,6 +158,50 @@ DOI = {10.5194/gmd-16-5601-2023}
       url={https://arxiv.org/abs/2501.18680},
 }
 
+@ARTICLE{van_lav_2025,
+       author = {{van Buchem}, C.~P.~A. and {Zilinskas}, M. and {Miguel}, Y. and {van Westrenen}, W.},
+        title = "{LavAtmos 2.0: Incorporating volatile species in vaporisation models}",
+      journal = {A\&A},
+     keywords = {planets and satellites: atmospheres, planets and satellites: composition, planets and satellites: interiors, planets and satellites: terrestrial planets, Astrophysics - Earth and Planetary Astrophysics, Astrophysics - Instrumentation and Methods for Astrophysics},
+         year = 2025,
+        month = mar,
+       volume = {695},
+          eid = {A154},
+        pages = {A154},
+          doi = {10.1051/0004-6361/202450992},
+archivePrefix = {arXiv},
+       eprint = {2408.10863},
+ primaryClass = {astro-ph.EP},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2025A&A...695A.154V},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+
+
+@misc{zilinskas_canc_2025,
+      title={Characterising the Atmosphere of 55 Cancri e: 1D Forward Model Grid for Current and Future JWST Observations},
+      author={Mantas Zilinskas and Christiaan van Buchem and Sebastian Zieba and Yamila Miguel and Emily Sandford and Renyu Hu and Jayshil Patel and Aaron Bello-Arufe and Leoni Janssen and Shang-Min Tsai and Diana Dragomir and Michael Zhang},
+      year={2025},
+      eprint={2503.15844},
+      archivePrefix={arXiv},
+      primaryClass={astro-ph.EP},
+      doi={https://doi.org/10.1051/0004-6361/202554062},
+      url={https://arxiv.org/abs/2503.15844},
+}
+
+@article{ seidler_impact_2024,
+	author = {{Seidler, Fabian L.} and {Sossi, Paolo A.} and {Grimm, Simon L.}},
+	title = {Impact of oxygen fugacity on the atmospheric structure and emission spectra of ultra-hot rocky exoplanets},
+	DOI= "10.1051/0004-6361/202450546",
+	url= "https://doi.org/10.1051/0004-6361/202450546",
+	journal = {A\&A},
+	year = 2024,
+	volume = 691,
+	pages = "A159",
+}
+
+
+
 @ARTICLE{hu_cancri_2024,
        author = {{Hu}, Renyu and {Bello-Arufe}, Aaron and {Zhang}, Michael and {Paragas}, Kimberly and {Zilinskas}, Mantas and {van Buchem}, Christiaan and {Bess}, Michael and {Patel}, Jayshil and {Ito}, Yuichi and {Damiano}, Mario and {Scheucher}, Markus and {Oza}, Apurva V. and {Knutson}, Heather A. and {Miguel}, Yamila and {Dragomir}, Diana and {Brandeker}, Alexis and {Demory}, Brice-Olivier},
         title = "{A secondary atmosphere on the rocky exoplanet 55 Cancri e}",
