+++ docs/paper/paper.md
@@ -43,7 +43,7 @@ It is thought that all rocky planets go through a 'magma ocean' stage, where the
 
 Several theoretical studies have modelled the atmospheres and evolution of these young planets, but all have made several simplifying assumptions. @lichtenberg_vertically_2021 coupled a simple atmosphere climate model with an interior code to simulate magma ocean evolution, but they did not account for the possibility of convective stability in their atmospheres. @krissansen_was_2021 and @krissansen_erosion_2024 made similar assumptions. @selsis_cool_2023 used a pure-steam radiative-convective model to investigate convective stability, finding that it is likely to occur, but did not extend their work to coupled scenarios which explore the secular evolution of these planets or mixtures of gas species; it is unlikely that these atmospheres are exclusively composed of steam [@nicholls_redox_2024]. @piette_rocky_2023 similarly explored potential atmospheres on observable lava planets, but did not consider the physics of atmosphere-interior coupling and chose semi-arbitrary gas compositions. @zilinskas_observability_2023 used the HELIOS code to model potential rock-vapour envelopes of sub-Neptune and super-Earth exoplanets, finding that the opacity of various gaseous species (notably SiO) plays a key role in determining their structure and observable properties. The demand for realistic modelling in the context of secular magma ocean evolution is apparent.
 
-Ensuring sufficient spectral resolution is important in modelling the blanketing effect of these atmospheres, as resolving the opacity (and transparency) of their many gaseous components  is known to be key in setting the rate at which these planets can cool [@pierrehumbert_book_2010; @nicholls_redox_2024]. It is also important that we are able to run grids of models that explore the range of possible (and as-yet poorly constrained) conditions that these planets could exhibit, which demands efficient modelling given finite computational resources. Magma ocean crystallisation could take up to several Gyr in the presence of continuous tidal forcing and atmospheric blanketing [@walterova_thermal_2020; @driscoll_tidal_2015]. The numerical efficiency afforded by AGNI enables simulations of rocky planets over geological timescales as part of a coupled interior-atmosphere planetary evolution framework. AGNI has so far been used in @hammond_photometric_2024, @nicholls_convection_2025, @nicholls_tidal_2025, and @nicholls_escape_2025.
+Ensuring sufficient spectral resolution is important in modelling the blanketing effect of these atmospheres, as resolving the opacity of their many gaseous components  is known to be key in setting the rate at which these planets can cool [@pierrehumbert_book_2010; @nicholls_redox_2024]. It is also important that we are able to run grids of models that explore the range of possible (and as-yet poorly constrained) conditions that these planets could exhibit, which demands efficient modelling given finite computational resources. Magma ocean crystallisation could take up to several Gyr in the presence of continuous tidal forcing and atmospheric blanketing [@walterova_thermal_2020; @driscoll_tidal_2015]. The numerical efficiency afforded by AGNI enables simulations of rocky planets over geological timescales as part of a coupled interior-atmosphere planetary evolution framework. AGNI has so far been used in @hammond_photometric_2024, @nicholls_convection_2025, @nicholls_tidal_2025, and @nicholls_escape_2025.
 
 # Comparison with other codes
 
@@ -54,7 +54,7 @@ AGNI is developed with the view of being coupled into the PROTEUS framework[^2] 
 * solve for an atmospheric temperature structure that conserves energy and allows for convectively stable regions,
 * operate with sufficient speed such that it may participate in the exploration of a wide parameter space.
 
[…]
