### 2025-05-08 #7726 @nichollsh
Hi @MartianColonist and @arjunsavel,

Thanks again for your feedback. Ryan raised an issue in the AGNI repo related to the build workflow for SOCRATES, which AGNI depends upon. This has been resolved by adding some safety checks to the build scripts (notably for when the required libraries are not installed) and improving the documentation.

I have also updated the paper following your comments. The typos have been fixed and the formatting issues in the figure corrected. A comparison of AGNI to other codes (HELIOS, Exo_k) has been expanded in the "comparison with other codes" section. An example of similar tools has been included in a new section, highlighting the recent usage of HELIOS+LavAtmos to calculate synthetic observations of 55 Cancri e.

-Harrison
