### 2025-05-09 #7726 @MartianColonist
I have successfully installed AGNI and verified that the two tutorials run as described in the documentation. This concludes my code and documentation review.

One final question for @nichollsh: does AGNI currently support parallel computing? I noticed that only a single CPU was active during the tutorials, but I wasn't sure whether this is just a limitation from running in a Jupyter notebook. If AGNI supports multi-CPU parallel computing, that would be good to mention in the documentation and paper (or mention as a future development).

@arfon both myself and @arjunsavel have checked all items in our reviews. You should now be able to proceed with final editorial checks.
