### 2025-05-09 #7726 @nichollsh
Great to hear!

At the moment AGNI does not support parallel computing. There's an open issue for this (https://github.com/nichollsh/AGNI/issues/96). My main use-case for AGNI is part of the coupled PROTEUS code, usually running grids of simulations, so I parallelise the model in the sense that one simulation is run per core. This would be a nice improvement for when running AGNI standalone (as, for example, some BSc students are doing at the moment).

I will add this to the "Future developments" section of the paper.

Thanks again for all of your inputs on my code and my paper.
