### 2025-05-12 #7726 @nichollsh
Thanks, @arfon.
I have made a release and published it on Zenodo here: https://doi.org/10.5281/zenodo.15386790
