### 2025-05-20 #7726 @nichollsh
Hi @arfon, do you have any updates on this? I think the paper is ready for editorial input :)
