+++ src/rfm.jl
@@ -1,0 +1,228 @@
+# Contains module and functions for performing radiative transfer with RFM
+
+# Not for direct execution
+if (abspath(PROGRAM_FILE) == @__FILE__)
+    thisfile = @__FILE__
+    error("The file '$thisfile' is not for direct execution")
+end
+
+module rfm
+
+    # Import modules
+    using LoggingExtras
+
+    # Local files
+    import ..atmosphere
+
+    # Species supported by RFM (https://eodg.atm.ox.ac.uk/RFM/rfm_spec.html#gaslist)
+    const RFM_GASES::Array{String} = [
+        # HITRAN
+        "H2O", "CO2", "O3", "N2O", "CO", "CH4", "O2", "NO", "SO2", "NO2", "NH3", "HNO3",
+        "OH", "HF", "HCl", "HBr", "HI", "ClO", "OCS", "H2CO", "HOCl", "N2", "HCN", "CH3Cl",
+        "H2O2", "C2H2", "C2H6", "PH3", "COF2", "SF6", "H2S", "HCOOH", "HO2", "O", "ClONO2",
+        "HOBr", "C2H4", "CH3OH", "CH3Br", "CH3CN", "CF4", "C4H2", "HC3N", "H2", "CS",
+        "SO3", "C2N2", "COCl2", "SO", "CH3F", "GeH4", "CS2", "CH3I", "NF3", "C3H4", "CH3",
+        # Extra
+        "ClONO2", "N2O5", "SF6", "CCl4", "HNO4", "SF5CF3", "BrONO2", "ClOOCl", "CH3OH",
+        ]
+
+    """
+    **Write atmospheric profile for RFM from current state.**
+
+    Parameters:
+    - `atmos::atmosphere.Atmos_t`       atmosphere data struct
+    """
+    function write_profile(atmos::atmosphere.Atmos_t)
+        @debug "Write profile for RFM"
+
+        # Convert arrays
+        t_arr::Array{Float64} = reverse(atmos.tmpl) # K
+        p_arr::Array{Float64} = reverse(atmos.pl) / 1e5  * 1e3 # mbar
+        z_arr::Array{Float64} = reverse(atmos.zl) / 1e3 # km
+
+        # file counters
+        ir::Int = 0
+        nr::Int = 5
+
+        # header
+        outstr::String =  ""
+        outstr *= "         %d ! Profile Levels\n"%nlev
+
+        # height profile
+        outstr *= "*HGT [km]\n"
+        for i in 1:atmos.nlev_l
+            # line break
+            if ir >= nr
+                ir=1
+                outstr *= "\n"
+            else
+                ir += 1
+            end
+            # write next value
+            outstr *= @sprintf(" %11.7f",z_arr[i])
+        end
+
+        # pressure profile
+        ir=0
+        outstr *="\n"
+        outstr *= "*PRE [mb]\n"
+        for i in 1:atmos.nlev_l
+            if ir >= nr
+                ir=1
+                outstr *= "\n"
+            else
+                ir += 1
+            end
+            outstr *= @sprintf(" %.6e",p_arr[i])
+        end
+
+        # temperature profile
+        ir=0
+        outstr *= "\n"
+        outstr *= "*TEM [K]\n"
+        for i in 1:atmos.nlev_l
+            if ir >= nr
+                ir=1
+                outstr*="\n"
+            else
+                ir += 1
+            end
+            outstr *= @sprintf(" %8.3f", t_arr[i])
+        end
+
+
+        # gases
+        for g in atmos.gas_names
+
+            # get mixing ratio and extend
+            x_arr = reverse(atmos.gas_vmr[g])
+            push!(x_arr, x_arr[end])
+
+            # clamp range
+            clamp!(x_arr, 0.0, 1.0)
+
+            # convert units to ppmv
[…]
