### 2025-05-21 95f9d59
Read RFM output and save to NetCDF with other data. Info in docs. par_files folder
