### 2025-05-21 f2618e7
Read RFM output and save to NetCDF with other data. Info in docs. par_files folder
