### 2025-05-21 2c79db0
Add comparison notebook
