### 2025-05-21 3dfb8cd
Update notebook. version 1.6.0
