### 2025-05-21 0cbd0ec
Update notebook. version 1.6.0
