### 2025-05-22 64ac2f0
Script for getting parfiles from OSF. Doc updates. Validation in config parser
