### 2025-05-22 c2332b2
Script for getting parfiles from OSF. Doc updates. Validation in config parser
