+++ res/par_files/_readme.txt
@@ -1,3 +1,0 @@
-These files contain linelist data tabulated in the HITRAN 'par' format.
-
-You must place them here manually, for now.
