+++ docs/src/setup.md
@@ -55,13 +55,3 @@ Of course, it is first necessary to set up FastChem, which can be done by runnin
 ```
 You **must** then set the `FC_DIR` environment variable to the location of the FastChem
 installation folder. Ideally you should also add this variable to your bashrc file.
-
-## Coupling with RFM
-The RFM radiative transfer code is proprietary. If you have obtained the code, coupling with
-AGNI can be enabled by setting the `RFM_DIR` environment variable to the location of the
-RFM root directory. 
-
-You should then set configuration parameter `execution.rfm_parfile` to the
-desired HITRAN-formatted (.par) linelist on your computer. Obtain these par files
-from the HITRAN website: https://hitran.org/lbl/
-
