+++ res/parfiles/_readme.txt
@@ -1,0 +1,5 @@
+These files contain linelist data tabulated in the HITRAN 'par' format.
+
+You can obtain examples with `src/get_data.sh parfiles`.
+
+Other files can be obtained from the HITRAN website: https://hitran.org/lbl/
