+++ .github/workflows/documentation.yml
@@ -29,6 +29,11 @@ jobs:
     steps:
       - uses: actions/checkout@v4
 
+      - name: NetCDF
+        run: |
+          sudo apt update
+          sudo apt-get install libnetcdff-dev netcdf-bin gfortran gcc
+
       - uses: julia-actions/setup-julia@v2
         with:
           version: '1.10'
