### 2025-05-26 d4b7bd6
Fix path issue for anim-frames. Fix import for median()
