+++ res/config/hotdry.toml
@@ -16,7 +16,7 @@ title = "Hot and dry"
     wind_speed      = 10.0
 
 [files]
-    input_sf        = "res/spectral_files/Frostflow/256/Frostflow.sf"
+    input_sf        = "res/spectral_files/Frostflow/16/Frostflow.sf"
     input_star      = "res/stellar_spectra/sun.txt"
     output_dir      = "out/"
 
