### 2025-05-26 22ba80f
Fix path issue for anim-frames. Fix import for median()
