### 2025-05-26 e355932
Plot for flux divergence. Account for input flux divergence in solver
