+++ docs/paper/paper.bib
@@ -19,7 +19,7 @@ volume = {122},
 number = {531},
 pages = {689-719},
 keywords = {Climate modelling, Cloud radiative properties, Numerical weather prediction, Portable code, Radiation parametrization, Variable spectral resolution},
-doi = {https://doi.org/10.1002/qj.49712253107},
+doi = {10.1002/qj.49712253107},
 url = {https://rmets.onlinelibrary.wiley.com/doi/abs/10.1002/qj.49712253107},
 eprint = {https://rmets.onlinelibrary.wiley.com/doi/pdf/10.1002/qj.49712253107},
 abstract = {Abstract A comprehensive new radiation code based on the two-stream equations in both the long-wave and short-wave spectral regions is described. The spectral resolution of the code is variable, enabling it to be used in a wide range of applications. Because of its flexibility, the code is well-suited to the investigation of the sensitivity of radiative calculations to changes in the way in which physical processes are parametrized. The gaseous transmission data are derived from a line-by-line model. Particular attention is directed towards the treatment of the water vapour continuum, the overlap between gases, and the sensitivity to changing the carbon dioxide concentrations. The performance of the code is examined both at high spectral resolution and in a lower-resolution configuration designed for the UK Meteorological Office Unified Forecast/Climate Model (UM). Particularly for use in the UM, the code must be shown to perform satisfactorily across the whole range of atmospheric conditions. Comparisons are therefore made with reference calculations in both the long-wave and the short-wave, in clear and cloudy skies, and the accuracy with which various processes may be represented is studied. For the cloudy calculations in the short-wave, a new method is presented for deriving the single-scattering properties in broad bands, based on the analytic expression for the reflectivity of an optically thick cloud. This minimizes the errors in calculating the short-wave radiative properties of water clouds when the spectral resolution is reduced to that designed for the UM. In contrast, for ice clouds the errors are minimized by deriving the single-scattering properties using linear averaging, as appropriate for optically thin clouds. In the long-wave, the vertical distribution of the radiative heating in cirrus clouds is examined at high spectral resolution. The effect of scattering of long-wave radiation, usually ignored in large-scale models, is examined in some detail and is explained using a simple model. Taking all these studies into account, it is concluded that the configuration designed for the UM retains the generality of the code, without significantly compromising the overall accuracy.},
@@ -185,7 +185,7 @@ archivePrefix = {arXiv},
       eprint={2503.15844},
       archivePrefix={arXiv},
       primaryClass={astro-ph.EP},
-      doi={https://doi.org/10.1051/0004-6361/202554062},
+      doi={0004-6361/202554062},
       url={https://arxiv.org/abs/2503.15844},
 }
 
@@ -460,7 +460,7 @@ volume = {129},
 number = {12},
 pages = {e2024JE008576},
 keywords = {magma oceans, lava planets, exoplanets, atmospheres, simulation, convection},
-doi = {https://doi.org/10.1029/2024JE008576},
+doi = {2024JE008576},
 url = {https://agupubs.onlinelibrary.wiley.com/doi/abs/10.1029/2024JE008576},
 eprint = {https://agupubs.onlinelibrary.wiley.com/doi/pdf/10.1029/2024JE008576},
 note = {e2024JE008576 2024JE008576},
