### 2025-05-26 #7726 @nichollsh
Thanks, @arfon.
I've updated the three entries above to remove the `https://doi.org/` component. Should be fixed now?
