+++ docs/paper/paper.bib
@@ -226,6 +226,7 @@ archivePrefix = {arXiv},
 	publisher = {Cambridge University Press},
 	author = {Pierrehumbert, Raymond},
 	year = {2010},
+	doi = {10.1017/CBO9780511780783}
 }
 
 @article{lacis_corrk_1991,
@@ -498,10 +499,10 @@ year = {2024}
 
 @article{nicholls_escape_2025,
        author = {Harrison Nicholls and Tim Lichtenberg and Richard Chatterjee and Emma Postolec and Raymond Pierrehumbert},
-        title = "A sub-Neptunian origin of the super-Earth L 98-59 d",
-      journal = {In preparation},
+        title = "Volatile-rich evolution of molten super-Earth L 98-59 d",
+      journal = {Submitted},
          year = 2025,
-         month = {June}
+         month = {May}
 }
 
 @ARTICLE{lichtenberg_review_2024,
@@ -585,7 +586,12 @@ journal = {The Astrophysical Journal},
 abstract = {Super-Earths span a wide range of bulk densities, indicating a diversity in interior conditions beyond that seen in the solar system. In particular, an emerging population of low-density super-Earths may be explained by volatile-rich interiors. Among these, low-density lava worlds have dayside temperatures that are high enough to evaporate their surfaces, providing a unique opportunity to probe their interior compositions and test for the presence of volatiles. In this work, we investigate the atmospheric observability of low-density lava worlds. We use a radiative-convective model to explore the atmospheric structures and emission spectra of these planets, focusing on three case studies with high observability metrics and substellar temperatures spanning ∼1900–2800 K: HD 86226 c, HD 3167 b, and 55 Cnc e. Given the possibility of mixed volatile and silicate interior compositions for these planets, we consider a range of mixed volatile and rock-vapor atmospheric compositions. This includes a range of volatile fractions and three volatile compositions: water-rich (100% H2O), water with CO2 (80% H2O+20% CO2), and a desiccated O-rich scenario (67% O2+33% CO2). We find that spectral features due to H2O, CO2, SiO, and SiO2 are present in the infrared emission spectra as either emission or absorption features, depending on dayside temperature, volatile fraction, and volatile composition. We further simulate JWST secondary-eclipse observations for each of the three case studies, finding that H2O and/or CO2 could be detected with as few as ∼five eclipses. Detecting volatiles in these atmospheres would provide crucial independent evidence that volatile-rich interiors exist among the super-Earth population.}
 }
 
-@book{perryman_book_2018, place={Cambridge}, edition={2}, title={The Exoplanet Handbook}, publisher={Cambridge University Press}, author={Perryman, Michael}, year={2018}} <div></div>
+@book{perryman_book_2018,
+place={Cambridge}, edition={2}, title={The Exoplanet Handbook},
+publisher={Cambridge University Press},
+author={Perryman, Michael}, year={2018},
+doi={10.1017/9781108304160}
+}
 
 @article{ortenzi_redox_2020,
 author={Ortenzi, G. and Noack, L. and Sohl, F. and Guimond, C. M. and Grenfell, J. L. and Dorn, C. and Schmidt, J. M. and Vulpius, S. and Katyal, N. and Kitzmann, D. and Rauer, H.},
