+++ docs/paper/paper.bib
@@ -178,16 +178,17 @@ archivePrefix = {arXiv},
 
 
 
-@misc{zilinskas_canc_2025,
-      title={Characterising the Atmosphere of 55 Cancri e: 1D Forward Model Grid for Current and Future JWST Observations},
-      author={Mantas Zilinskas and Christiaan van Buchem and Sebastian Zieba and Yamila Miguel and Emily Sandford and Renyu Hu and Jayshil Patel and Aaron Bello-Arufe and Leoni Janssen and Shang-Min Tsai and Diana Dragomir and Michael Zhang},
-      year={2025},
-      eprint={2503.15844},
-      archivePrefix={arXiv},
-      primaryClass={astro-ph.EP},
-      doi={0004-6361/202554062},
-      url={https://arxiv.org/abs/2503.15844},
+@article{zilinskas_canc_2025,
+	author = {{Zilinskas, M.} and {van Buchem, C. P. A.} and {Zieba, S.} and {Miguel, Y.} and {Sandford, E.} and {Hu, R.} and {Patel, J. A.} and {Bello-Arufe, A.} and {Janssen, L. J.} and {Tsai, S.-M.} and {Dragomir, D.} and {Zhang, M.}},
+	title = {Characterising the atmosphere of 55 Cancri e - 1D forward model grid for current and future JWST observations},
+	DOI= "10.1051/0004-6361/202554062",
+	url= "https://doi.org/10.1051/0004-6361/202554062",
+	journal = {A\&A},
+	year = 2025,
+	volume = 697,
+	pages = "A34",
 }
+
 
 @article{ seidler_impact_2024,
 	author = {{Seidler, Fabian L.} and {Sossi, Paolo A.} and {Grimm, Simon L.}},
@@ -461,7 +462,7 @@ volume = {129},
 number = {12},
 pages = {e2024JE008576},
 keywords = {magma oceans, lava planets, exoplanets, atmospheres, simulation, convection},
-doi = {2024JE008576},
+doi = {10.1029/2024JE008576},
 url = {https://agupubs.onlinelibrary.wiley.com/doi/abs/10.1029/2024JE008576},
 eprint = {https://agupubs.onlinelibrary.wiley.com/doi/pdf/10.1029/2024JE008576},
 note = {e2024JE008576 2024JE008576},
