### 2025-05-29 e1ff241
Removed duplicated bib entry
