+++ docs/paper/paper.bib
@@ -572,21 +572,6 @@ title = {Redox States of Initial Atmospheres Outgassed on Rocky Planets and Plan
 journal = {The Astrophysical Journal},
 }
 
-@article{piette_rocky_2023,
-doi = {10.3847/1538-4357/acdef2},
-url = {https://dx.doi.org/10.3847/1538-4357/acdef2},
-year = {2023},
-month = {aug},
-publisher = {The American Astronomical Society},
-volume = {954},
-number = {1},
-pages = {29},
-author = {Anjali A. A. Piette and Peter Gao and Kara Brugman and Anat Shahar and Tim Lichtenberg and Francesca Miozzi and Peter Driscoll},
-title = {Rocky Planet or Water World? Observability of Low-density Lava World Atmospheres},
-journal = {The Astrophysical Journal},
-abstract = {Super-Earths span a wide range of bulk densities, indicating a diversity in interior conditions beyond that seen in the solar system. In particular, an emerging population of low-density super-Earths may be explained by volatile-rich interiors. Among these, low-density lava worlds have dayside temperatures that are high enough to evaporate their surfaces, providing a unique opportunity to probe their interior compositions and test for the presence of volatiles. In this work, we investigate the atmospheric observability of low-density lava worlds. We use a radiative-convective model to explore the atmospheric structures and emission spectra of these planets, focusing on three case studies with high observability metrics and substellar temperatures spanning ∼1900–2800 K: HD 86226 c, HD 3167 b, and 55 Cnc e. Given the possibility of mixed volatile and silicate interior compositions for these planets, we consider a range of mixed volatile and rock-vapor atmospheric compositions. This includes a range of volatile fractions and three volatile compositions: water-rich (100% H2O), water with CO2 (80% H2O+20% CO2), and a desiccated O-rich scenario (67% O2+33% CO2). We find that spectral features due to H2O, CO2, SiO, and SiO2 are present in the infrared emission spectra as either emission or absorption features, depending on dayside temperature, volatile fraction, and volatile composition. We further simulate JWST secondary-eclipse observations for each of the three case studies, finding that H2O and/or CO2 could be detected with as few as ∼five eclipses. Detecting volatiles in these atmospheres would provide crucial independent evidence that volatile-rich interiors exist among the super-Earth population.}
-}
-
 @book{perryman_book_2018,
 place={Cambridge}, edition={2}, title={The Exoplanet Handbook},
 publisher={Cambridge University Press},
