### 2025-05-29 #7726 @nichollsh
This error seems to have been caused by a duplicated reference introduced during the revisions. @arfon, I have removed the duplication and the proof seems to compile fine. Does this fix the above XML error?
