### 2025-05-30 #7726 @editorialbot
:wave: @openjournals/aass-eics, this paper is ready to be accepted and published.
