### 2025-05-30 fe5e6b5
Add JOSS citation
