### 2025-06-02 fa05c14
Editorial tweaks for JOSS paper
