+++ docs/paper/paper.md
@@ -33,13 +33,11 @@ bibliography: paper.bib
 
 It is important that we are able to accurately model the atmospheres of (exo)planets. This is because atmospheres play a central role in setting a planet's thermochemical environment at a given point in time, and also in regulating how it evolves over geological timescales. Additionally, it is primarily by observation of their atmospheres that we are able to characterise exoplanets. There is particular demand for accurate models in the context of so-called lava worlds: planets with molten interiors (or 'magma oceans').
 
-AGNI is a Julia program designed to solve for the temperature and radiation environment within the atmospheres of rocky (exo)planets. It leverages a well established FORTRAN code [@edwards_studies_1996; @sergeev_socrates_2023] to calculate radiative fluxes from a given atmospheric temperature structure and composition, which -- alongside representations of convection and other processes -- enables an energy-conserving numerical solution for the atmospheric conditions. In contrast to most other numerical atmosphere models, AGNI uses a Newton-Raphson optimisation method to obtain its solution, which enables improved performance and scalability. Our model was specifically developed for use alongside planetary interior models within a coupled simulation framework. However, it can also be applied to scientific problems standalone when used as an executable program; it reads TOML configuration files and outputs figures and NetCDF datasets. AGNI can also function as a software library; it is used in this sense within the [Jupyter notebook tutorials](https://github.com/nichollsh/AGNI/tree/main/tutorials) of our GitHub repository[^1].
-
-[^1]: [nichollsh.github.io/AGNI](https://nichollsh.github.io/AGNI/dev/)
+AGNI is a Julia program designed to solve for the temperature and radiation environment within the atmospheres of rocky (exo)planets. It leverages a well established FORTRAN code [@edwards_studies_1996; @sergeev_socrates_2023] to calculate radiative fluxes from a given atmospheric temperature structure and composition, which -- alongside representations of convection and other processes -- enables an energy-conserving numerical solution for the atmospheric conditions. In contrast to most other numerical atmosphere models, AGNI uses a Newton--Raphson optimisation method to obtain its solution, which enables improved performance and scalability. Our model was specifically developed for use alongside planetary interior models within a coupled simulation framework. However, it can also be applied to scientific problems standalone when used as an executable program; it reads TOML configuration files and outputs figures and NetCDF datasets. AGNI can also function as a software library; it is used in this sense within the [Jupyter notebook tutorials](https://github.com/nichollsh/AGNI/tree/main/tutorials) of our [GitHub repository](https://nichollsh.github.io/AGNI/dev/).
 
 # Statement of need
 
[…]
